% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_additional_text.R
\name{add_additional_text}
\alias{add_additional_text}
\title{add_additional_text()}
\usage{
add_additional_text(wb, text, sheetname, start_row)
}
\arguments{
\item{wb}{Workbook object where text should be inserted}

\item{text}{Character vector of text to be inserted}

\item{sheetname}{Name of the worksheet where text should be inserted.}

\item{start_row}{Row from which to insert elements}
}
\value{
Integer corresponding to the final row index
}
\description{
Insert text into an existing workbook object. Text is inserted
 at column index 1.
}
\keyword{internal}
