% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_sheetname}
\alias{add_sheetname}
\alias{add_title}
\alias{add_source}
\alias{add_metadata}
\alias{add_grouplines}
\alias{add_group_names}
\alias{add_plot_height}
\alias{add_plot_width}
\alias{add_plot_size}
\title{Helper functions for adding attributes to input objects}
\usage{
add_sheetname(object, value)

add_title(object, value)

add_source(object, value, prefix = NULL, collapse = NULL)

add_metadata(object, value, prefix = NULL, collapse = NULL)

add_grouplines(object, value)

add_group_names(object, value)

add_plot_height(object, value)

add_plot_width(object, value)

add_plot_size(object, value)
}
\arguments{
\item{object}{The object to add an attribute to}

\item{value}{A value}

\item{prefix}{A prefix, default to NULL}

\item{collapse}{Separator to collapse character vectors on, defaults to NULL}
}
\description{
A set of helper functions which serve the purpose of allowing
  users to assign information like sheetnames, titles, sources, metadata, and
  more to an object. These attributes are then used in place of the function
  arguments.
}
\details{
\code{add_sheetname}: Adds a sheetname.

  \code{add_title}: Adds a title

  \code{add_source}: Adds a source. Behavior can be modified further via the
    prefix and collapse argument. Setting collapse to a value other than NULL
    will result in the input provided in value being concatenated.

  \code{add_metadata}: Adds metadata. Behavior can be modified further via the
    prefix and collapse argument, just like with \code{add_source}.

  \code{add_grouplines}: Adds group lines. Throws an error if input object is
    not a data.frame.

  \code{add_group_names}: Adds group names. Throws an error if group lines
    have not been set.

  \code{add_plot_height}: Adds plot height. Throws an error if input object is
    not a valid plot type.

  \code{add_plot_width}: Adds plot width. Throws an error if input object is
    not a valid plot type.

  \code{add_plot_size}: Assigns both height and width in one call. Expects a
    vector of length 2, with the first element corresponding to width, and the
    second to height.
}
\examples{
library(dplyr)
library(ggplot2)

# Load data and assign attributes in one pipeline
df <- mtcars \%>\%
  add_sheetname("Cars") \%>\%
  add_title("Motor Trend Car Road Tests") \%>\%
  add_source(
    c("Henderson and Velleman (1981),",
      "Building multiple regression models interactively.",
      "Biometrics, 37, 391–411."),
    prefix = "Source:",
    collapse = " ") \%>\%
  add_metadata(
    c("The data was extracted from the 1974 Motor",
      "Trend US magazine and comprises fuel consumption",
      "and 10 aspects of automobile design and",
      "performance for 32 automobiles (1973–74 models)."),
    collapse = " ")

# Create a plot and assign attributes in one pipeline
plt <- (ggplot(mtcars) +
  geom_histogram(aes(x = hp))) \%>\%
    add_sheetname("PS") \%>\%
    add_title("Histogram of horsepower") \%>\%
    add_plot_size(c(6, 3))

\dontrun{
# Generate outputfile using minimal call
datasetsXLSX(
  file = tempfile(fileext = ".xlsx"),
  datasets = list(df, plt))
}
}
