% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_group_names}
\alias{add_group_names}
\title{add_group_names()}
\usage{
add_group_names(object, value)
}
\arguments{
\item{object}{The object to add an attribute to}

\item{value}{A value}
}
\description{
Add title to an object
}
