% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_grouplines}
\alias{add_grouplines}
\title{add_grouplines()}
\usage{
add_grouplines(object, value)
}
\arguments{
\item{object}{The object to add an attribute to}

\item{value}{A value}
}
\description{
Add title to an object
}
