% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_plot_height}
\alias{add_plot_height}
\title{add_plot_height()}
\usage{
add_plot_height(object, value)
}
\arguments{
\item{object}{The object to add an attribute to}

\item{value}{A value}
}
\description{
Add title to an object
}
