% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_source}
\alias{add_source}
\title{add_title()}
\usage{
add_source(object, value, prefix = NULL, collapse = NULL)
}
\arguments{
\item{object}{The object to add an attribute to}

\item{value}{A value}

\item{prefix}{A prefix, default to NULL}

\item{collapse}{Separator to collapse character vectors on, defaults to NULL}
}
\description{
Add title to an object
}
