% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_title}
\alias{add_title}
\title{add_title()}
\usage{
add_title(object, value)
}
\arguments{
\item{object}{The object to add an attribute to}

\item{value}{A value}
}
\description{
Add title to an object
}
