% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsXLSX_helpers.R
\name{addContentObject}
\alias{addContentObject}
\title{addContentObject()}
\usage{
addContentObject(
  list = list(),
  object,
  sheetname,
  title,
  sources,
  metadata = NA,
  grouplines = NA,
  group_names = NA,
  width = 5,
  height = 5
)
}
\value{
A numeric vector of row indices
}
\description{
Add an object, either a data.frame or a supported plot object
  to a list together with
}
