% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsXLSX_helpers.R
\name{addDataObject}
\alias{addDataObject}
\title{addDataObject()}
\usage{
addDataObject(
  list,
  object,
  sheetname,
  title,
  sources,
  metadata,
  grouplines,
  group_names
)
}
\value{
A list
}
\description{
Add a data.frame with metadata to a list for use in datasetsXLSX
}
\keyword{internal}
