% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsXLSX_helpers.R
\name{addPlotObject}
\alias{addPlotObject}
\title{namedRegionRowExtent()}
\usage{
addPlotObject(list, object, sheetname, width, height)
}
\value{
A numeric vector of row indices
}
\description{
Get row extent of a named region in a workbook object
}
\keyword{internal}
