% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{addUserConfig}
\alias{addUserConfig}
\title{Add a new configuration using a name and path}
\usage{
addUserConfig(name = "default", path = NULL, store_path = "~/.config/R/statR")
}
\description{
Add a new configuration using a name and path
}
