% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aXLSX.R
\name{aXLSX}
\alias{aXLSX}
\title{aXLSX()}
\usage{
aXLSX(
  data,
  file,
  title = "Title",
  source = "statzh",
  metadata = NA,
  logo = "statzh",
  grouplines = NA,
  contactdetails = "statzh",
  author = "user"
)
}
\arguments{
\item{data}{data to be exported.}

\item{file}{file name of the xlsx-file. The extension ".xlsx" is added automatically.}

\item{title}{title to be put above the data in the worksheet.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata-information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"}

\item{grouplines}{columns to be separated visually by vertical lines.}

\item{contactdetails}{contactdetails of the data publisher. Defaults to "statzh".}

\item{author}{defaults to last two letters (initials) or numbers of the user name.}
}
\description{
Function to export data from R to a formatted .xlsx-file. The
 data is exported to the first sheet. Metadata information is exported to
 the second sheet.
}
\examples{
dataset <- mtcars
source_string <- paste("Source: Henderson and Velleman (1981).",
  "Building multiple regression models interactively.",
  "Biometrics, 37, 391–411.")

metadata_string <- paste("The data was extracted from the 1974",
  "Motor Trend US magazine and comprises fuel consumption and",
  "10 aspects of automobile design and performance for 32 automobiles",
  "(1973–74 models).")
\donttest{
\dontrun{
aXLSX(data = mtcars,
      title = "Motor trend car road tests",
      file = "motor_trend_car_road_tests",
      source = source_string,
      metadata = metadata_string,
      contactdetails = "statzh",
      grouplines = NA,
      logo = "statzh",
      author = "user")
}
}
}
\keyword{aXLSX}
