% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{bridge}
\alias{bridge}
\title{Bridge Regression}
\usage{
bridge(X, y, lam, alpha, tol = 10^(-5), maxit = 1000)
}
\arguments{
\item{X}{matrix or data frame}

\item{y}{matrix or data frame of response values}

\item{lam}{tuning parameter for regularization term}

\item{alpha}{tuning parameter. alpha in (1, 2).}

\item{tol}{tolerance - used to determine algorithm convergence}

\item{maxit}{maximum iterations}
}
\value{
returns beta estimates, total iterations, and gradients.
}
\description{
Computes the bridge-penalized linear regression estimates. We assume alpha in (1, 2). Uses the MM algorithm to compute.
}
\examples{
library(dplyr)
X = dplyr::select(iris, -c(Species, Sepal.Length))
y = dplyr::select(iris, Sepal.Length)
Ridge(X, y, lam = 0.1, alpha = 1.5)
}

