% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{bridge_gradient}
\alias{bridge_gradient}
\title{Gradient of Bridge-Penalized Regression}
\usage{
bridge_gradient(XX, Xy, b, lam, alpha)
}
\arguments{
\item{XX}{crossprod(X, X) - X is centered, no intercept}

\item{Xy}{crossprod(X, y) - y is centered}

\item{b}{beta estimates (intercept not included)}

\item{lam}{tuning parameter for regularization term}

\item{alpha}{tuning parameter. Alpha is contained in (1, 2).}
}
\value{
returns the gradient.
}
\description{
Computes the gradient of bridge regression. We use this to determine if the KKT conditions are satisfied.
}
\examples{
bridge_gradient(XX, Xy, b = betas, lam = 0.1, alpha = 1.5)
}

