% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_bridge.R
\name{bridge_logistic}
\alias{bridge_logistic}
\title{Bridge-Penalized Logistic Regression}
\usage{
bridge_logistic(X, y, n.list, lam, alpha, tol = 10^(-5), maxit = 1e+05)
}
\arguments{
\item{X}{Matrix or data frame}

\item{y}{response vector of 0,1}

\item{n.list}{vector containing the number of replicates for each observation}

\item{lam}{tuning parameter for regularization term}

\item{alpha}{tuning parameter. Alpha is contained in (1, 2).}

\item{tol}{tolerance - used to determine algorithm convergence}

\item{maxit}{maximum iterations}
}
\value{
returns beta estimates (includes intercept), total iterations, and gradients.
}
\description{
Computes the bridge-penalized logistic regression estimates. We assume alpha is contained in (1, 2). Uses the MM algorithm to compute.
}
\examples{
bridge_logistic(X, y, n.list = c(rep(1, n)), lam = 0.1, alpha = 1.5)
}

