% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_search.R
\name{bsearch}
\alias{bsearch}
\title{Bisection search}
\usage{
bsearch(dg, a, b, L = 1e-07, quiet = FALSE)
}
\arguments{
\item{dg}{the derivative of the function to minimize, where dg(u, ...) is the function evaluated at u.}

\item{a}{left endpoint of the initial interval of uncertainty.}

\item{b}{right endpoint of the initial interval of uncertainty.}

\item{L}{the maximum length of the final interval of uncertainty.}

\item{quiet}{should the function stay quiet?}

\item{...}{additional argument specifications for dg}
}
\value{
returns the midpoint of the final interval of uncertainty.
}
\description{
Minimizes a univariate strictly pseudoconvex function over the interval [a, b]. This is augmented code from Adam Rothman's STAT 8054 course (2017).
}
\examples{
bsearch(dg, -10, 10, quiet = T)
}
