% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{check_for_attributes}
\alias{check_for_attributes}
\title{Check for presence of keyword attributes in input data.frame}
\usage{
check_for_attributes(object)
}
\arguments{
\item{object}{Input data.frame}
}
\description{
A function which checks for presence of certain attributes in an input
data.frame.
}
\keyword{internal}
