% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_input.R
\name{check_sheetname}
\alias{check_sheetname}
\title{Functions for checking and transforming user input}
\usage{
check_sheetname(sheetname)
}
\description{
Helper functions not intended to be called directly by users.
}
\keyword{internal}
