% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CV_logisticc}
\alias{CV_logisticc}
\title{CV Logisticc (c++)}
\usage{
CV_logisticc(X, y, lam = 0L, alpha = 0L, penalty = "none",
  intercept = TRUE, method = "IRLS", tol = 1e-05, maxit = 10000,
  vec = 0L, init = 0L, criteria = "logloss", K = 5L)
}
\arguments{
\item{X}{matrix}

\item{y}{matrix or vector of response values 0,1}

\item{lam}{vector of tuning parameters for ridge regularization term. Defaults to `lam = 0`}

\item{alpha}{vector of tuning parameters for bridge regularization term. Defaults to 'alpha = 1.5'}

\item{penalty}{choose from c('none', 'ridge', 'bridge'). Defaults to 'none'}

\item{intercept}{Defaults to TRUE}

\item{method}{optimization algorithm. Choose from 'IRLS' or 'MM'. Defaults to 'IRLS'}

\item{tol}{tolerance - used to determine algorithm convergence. Defaults to 1e-5}

\item{maxit}{maximum iterations. Defaults to 1e5}

\item{vec}{optional vector to specify which coefficients will be penalized}

\item{init}{optional initialization for MM algorithm}

\item{criteria}{specify the criteria for cross validation. Choose from c('mse', 'logloss', 'misclass'). Defauls to 'logloss'}

\item{K}{specify number of folds in cross validation, if necessary}
}
\value{
returns best lambda, best alpha, and cross validation errors
}
\description{
Computes the coefficient estimates for logistic regression. ridge regularization and bridge regularization optional. This function is to be used with the 'logisticc' function.
}
\examples{
CV_logisticc(X, y, lam = seq(0.1, 2, 0.1), alpha = c(1.1, 1.9, 0.1), penalty = 'bridge', method = 'MM', vec = c(0,1,1,1))

}
