% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{CVsplit}
\alias{CVsplit}
\title{CV split}
\usage{
CVsplit(X, Y, split = 0.5)
}
\arguments{
\item{X}{nxp data matrix. Each row corresponds to a single observation and each column contains n observations of a single feature/variable.}

\item{Y}{nxr response matrix. Each row corresponds to a single response and each column contains n response of a single feature/response.}

\item{split}{fraction of objects devoted to training set}
}
\value{
X.train, Y.train, X.test, Y.test
}
\description{
splits data objects into training and testing sets
}
