% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{Normal Data Generator}
\usage{
data_gen(n, p, theta, var = 0.5, reps = 200)
}
\arguments{
\item{n}{desired sample size}

\item{p}{desired dimension}

\item{theta}{blah}

\item{var}{blah}

\item{reps}{blah}
}
\description{
generates random data.
}
\examples{
data_gen(1000, 10, 0.5)
}

