% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsXLSX.R
\name{datasetsXLSX_plot}
\alias{datasetsXLSX_plot}
\title{datasetsXLSX_withplot()}
\usage{
datasetsXLSX_plot(
  file,
  datasets,
  sheetnames,
  titles,
  sources,
  plot_widths,
  plot_heights,
  metadata = NA,
  grouplines = NA,
  group_names = NA,
  overwrite = FALSE
)
}
\arguments{
\item{file}{file}

\item{datasets}{datasets or plots to be included.}

\item{sheetnames}{names of the sheet tabs.}

\item{titles}{titles of the different sheets.}

\item{sources}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata information to be included. Defaults to NA.}

\item{grouplines}{Column for second header(s). Format: List e.g list(c(2,4,6))}

\item{group_names}{Name(s) of the second header(s). Format: List e.g
list(c("title 1", "title 2", "title 3"))}

\item{overwrite}{overwrites the existing excel files with the same file name.
default to FALSE}
}
\description{
Function to export several datasets and/or figures from R to an
 .xlsx-file. The function creates an overview sheet and separate sheets for
 each dataset/figure.
}
\keyword{datasetsXLSX}
