% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{dense}
\alias{dense}
\title{Generate dense matrices}
\usage{
dense(p = 8, base = 0.9, n = 100, X = FALSE)
}
\arguments{
\item{p}{desired dimension}

\item{base}{base multiplier}
}
\description{
Generate p-dimensional matrices so that its inverse is dense.
}
\examples{
dense(p = 10, base = 0.9)
}
