% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_search.R
\name{derivative}
\alias{derivative}
\title{Derivative}
\usage{
derivative(g, x, delta = 1e-07)
}
\arguments{
\item{g}{the derivative of the function to minimize, where dg(u, ...) is the function evaluated at u.}

\item{x}{value to evaluate the derivative at}

\item{delta}{defaults to 10e-8}
}
\description{
Takes the approximate derivative for a given function
}
