% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{diagnostic}
\alias{diagnostic}
\title{Diagnostic}
\usage{
diagnostic(data., x., y.)
}
\arguments{
\item{data.}{data frame}

\item{x.}{x-axis}

\item{y.}{y-axis}
}
\value{
a residual plot and QQ plot
}
\description{
This function simply streamlines the process of creating diagnostic plots with ggplot
}
\examples{
diagnostic(iris, Sepal.Length, Sepal.Width)
}
