% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.statcol.all.R
\name{display.statcol.all}
\alias{display.statcol.all}
\title{display.statcol.all()}
\usage{
display.statcol.all()
}
\description{
Preview all Stat ZH palettes stored in the 'zhpal'-object
}
\examples{
display.statcol.all()

}
\keyword{display.statcol.all}
