% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_search.R
\name{dsearch}
\alias{dsearch}
\title{Dichotomous search}
\usage{
dsearch(g, a, b, L = 1e-07, eps = (L/2.1), quiet = FALSE)
}
\arguments{
\item{g}{the function to minimize, where g(u, ...) is the function evaluated at u.}

\item{a}{left endpoint of the initial interval of uncertainty.}

\item{b}{right endpoint of the initial interval of uncertainty.}

\item{L}{the maximum length of the final interval of uncertainty.}

\item{eps}{search parameter, must be less than L/2}

\item{quiet}{should the function stay quiet?}

\item{...}{additional argument specifications for g}
}
\value{
returns the midpoint of the final interval of uncertainty.
}
\description{
Minimizes a univariate strictly quasiconvex function over the interval [a, b]. This is augmented code from Adam Rothman's STAT 8054 course (2017).
}
\examples{
dsearch(g, -10, 10, quiet = T)
}

