% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{exact}
\alias{exact}
\title{Define statistics for a statistical procedure}
\usage{
exact(name, output = NULL, options = NULL)
}
\arguments{
\item{name}{The name of the statistical function}

\item{output}{The name of the output data frame}

\item{options}{The options to use for this procedure, wrapped in
an \code{\link{opts}} function.}
}
\value{
An "exact" object.
}
\description{
The \code{exact} function is used to define
statistics for a statistical procedure.
}
