% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{excelIndexToRowCol}
\alias{excelIndexToRowCol}
\title{Convert Excel index to row/column indices}
\usage{
excelIndexToRowCol(index)
}
\arguments{
\item{index}{A Microsoft Excel style index (see details)}
}
\value{
A list containing row indices and column indices
}
\description{
Converts an Excel style index (e.g. A1) into numeric row and
  column indices. Handles cells (A1) as well as matrices (A1:B2)
}
\keyword{internal}
