% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extract_attribute}
\alias{extract_attribute}
\title{extract_attribute()}
\usage{
extract_attribute(object, which, required_val = FALSE)
}
\arguments{
\item{object}{Object to extract attribute from. In practice a data.frame,
ggplot object, or a character string providing a path to an image.}

\item{which}{Which attribute to extract.}

\item{required_val}{Boolean, if TRUE tries to look up a default if attribute not
found, otherwise raises error}
}
\description{
Function to extract attributes from an object
}
\keyword{internal}
