% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extract_attributes}
\alias{extract_attributes}
\title{extract_attributes()}
\usage{
extract_attributes(datasets, which, required_val = FALSE)
}
\arguments{
\item{datasets}{List of objects}

\item{which}{Which attribute to extract.}

\item{required_val}{Boolean, if TRUE tries to look up a default if attribute not
found, otherwise raises error}
}
\description{
Function to extract attributes from multiple objects in a list
}
\keyword{internal}
