% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush_left.r
\name{flush_left}
\alias{flush_left}
\title{flush_left()}
\usage{
flush_left(g)
}
\arguments{
\item{g}{ggplot object}
}
\description{
Function to flush title, subtitle and caption to the
  lefthand side of the ggplot2 graphics device
}
\examples{
plt <- ggplot2::ggplot(mtcars, ggplot2::aes(x = cyl)) +
  ggplot2::geom_bar() +
  ggplot2::labs(title = "Cylinders", subtitle = "Bar chart",
                caption = "mtcars")

flush_left(plt)
}
\keyword{flush_left}
