% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Creates frequency statistics}
\usage{
freq(
  data,
  by = NULL,
  ...,
  tables = NULL,
  test = NULL,
  weight = NULL,
  options = NULL
)
}
\arguments{
\item{data}{The input data frame.}

\item{by}{The by variable for the procedure, wrapped in a \code{vars} function.}

\item{...}{A set of \code{\link{exact}} functions defining the statistical
procedures to use.}

\item{tables}{The variables to use for the tables parameter, wrapped in
a \code{\link{requests}} object.}

\item{test}{The tests to use for this procedure.}

\item{weight}{Which if any variables to weight the procedure.  Wrap in
a \code{vars} function if there is more than one.}

\item{options}{The options to use for this procedure, wrapped in an
\code{\link{opts}} function.}
}
\value{
A frequency object.
}
\description{
The \code{freq} function creates frequency statistics in a
manner similar to the SAS® procedure PROC FREQ.
}
\examples{
# freq (data=d1,
#       by=vars(v1, v2),
#       exact(fisher, fisherout, opts(ci)),
#       exact(chisq, chisqout),
#       tables = requests(v1, v2, v1 * v2),
#       test=list(),
#       weight=vars(cnt1),
#       options=opts()
# )
}
