% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fro}
\alias{fro}
\title{Mean Frobenius norm}
\usage{
fro(X)
}
\arguments{
\item{X}{object}
}
\value{
norm
}
\description{
calculates the average squared value of an object. That is, all elements are squared, summed, and divided by the total number of elements
}
