% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{get_user_config}
\alias{get_user_config}
\title{Extrahiert Defaults aus User Config und ersetzt Werte}
\usage{
get_user_config(config, params_to_check)
}
\arguments{
\item{config}{Name der User Config}

\item{params_to_check}{Vektor mit Parametern}
}
\description{
Extrahiert Defaults aus User Config und ersetzt Werte
}
\keyword{internal}
