% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_generator.R
\name{glm_generator}
\alias{glm_generator}
\title{GLM generator}
\usage{
glm_generator(x, data, family = gaussian)
}
\arguments{
\item{x}{string of formulas}

\item{data}{name of data object}

\item{family}{define type of GLM required}
}
\value{
nested tibble
}
\description{
Takes a string of formulas and trains GLM models to output
coefficients, diagnostic parameters and predicted values in a tibble.
}
\examples{
formulas <- c("mpg ~ hp", "mpg ~ cyl", "mpg ~ wt")
glm_generator(x = formulas, data = mtcars, family = gaussian)
}
\seealso{
\code{\link{statr}}
}
\author{
Jay Achar
}
