% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{gradient_IRLS_logistic}
\alias{gradient_IRLS_logistic}
\title{Gradient of Logistic Regression (IRLS)}
\usage{
gradient_IRLS_logistic(betas, X, y, lam = 0, vec)
}
\arguments{
\item{betas}{beta estimates (includes intercept)}

\item{X}{matrix or data frame}

\item{y}{response vector of 0,1}

\item{lam}{tuning parameter for ridge regularization term}

\item{vec}{vector to specify which coefficients will be penalized}
}
\value{
returns the gradient
}
\description{
Computes the gradient of logistic regression (optional ridge regularization term). We use this to determine if the KKT conditions are satisfied. This function is to be used with the 'IRLS' function.
}
\examples{
gradient_IRLS_logistic(betas, X, y, lam = 0.1, penalty = 'ridge')
}

