% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_bridge.R
\name{gradient_logistic_bridge}
\alias{gradient_logistic_bridge}
\title{Gradient of Bridge-Penalized Logistic Regression}
\usage{
gradient_logistic_bridge(X, y, N, b, lam, alpha)
}
\arguments{
\item{X}{matrix or data frame}

\item{y}{response vector of 0,1}

\item{N}{diag(n.list) - where n.list is a vector of the number of replicates for each observation}

\item{b}{beta estimates (includes intercept)}

\item{lam}{tuning parameter for regularization term}

\item{alpha}{tuning parameter. Alpha is contained in (1, 2).}
}
\value{
returns the gradient.
}
\description{
Computes the gradient of bridge-penalized logistic regression. We use this to determine if the KKT conditions are satisfied.
}
\examples{
logistic_bridge_gradient(X, y, N = diag(n.list), b = betas, lam = 0.1, alpha = 1.5)
}

