% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM.R
\name{gradient_MM_logistic}
\alias{gradient_MM_logistic}
\title{Gradient of Logistic Regression (MM)}
\usage{
gradient_MM_logistic(betas, X, y, lam = 0, alpha = 1.5, gamma = 1, vec)
}
\arguments{
\item{betas}{beta estimates (includes intercept)}

\item{X}{matrix or data frame}

\item{y}{response vector of 0,1}

\item{lam}{tuning parameter for ridge regularization term}

\item{alpha}{optional tuning parameter for bridge regularization term. Defaults to 'alpha = 1.5'}

\item{gamma}{indicator function. 'gamma = 1' for ridge, 'gamma = 0' for bridge. Defaults to 'gamma = 1'}

\item{vec}{vector to specify which coefficients will be penalized}
}
\value{
returns the gradient
}
\description{
Computes the gradient of logistic regression (optional ridge regularization term). We use this to determine if the KKT conditions are satisfied. This function is to be used with the 'MM' function.
}
\examples{
gradient_MM_logistic(betas, X, y, lam = 0.1, alpha = 1.5, penalty = 'bridge')
}

