% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{initUserConfigStore}
\alias{initUserConfigStore}
\title{Initialisiert die Excel-Header-Konfiguration}
\usage{
initUserConfigStore(store_path = "~/.config/R/statR")
}
\arguments{
\item{store_path}{Pfad unter welchem das Konfigurationsfile fuer das registrieren
von Header-Config-Yaml-Files liegen soll}
}
\description{
Die Header-Konfiguration wird ueber globale Options geloest, welche in
einem Yaml file definiert werden. Per default wird die Konfiguration des
Packages angezogen. Es besteht die Moeglichkeit, eine eigene Konfiguration
zu hinterlegen. Siehe dazu die Funktion: addUserConfig()
}
