% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{inputHelperContactInfo}
\alias{inputHelperContactInfo}
\title{Construct a vector with contact information from defaults}
\usage{
inputHelperContactInfo(compact = FALSE)
}
\arguments{
\item{compact}{A boolean. If TRUE, a shortened version is displayed}
}
\value{
A character vector
}
\description{
A helper function which constructs a character vector from the contact
information defined in the user profile. If \code{compact = TRUE}, the
organization is dropped, and the name and phone number are displayed in the
same row.
}
