% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{inputHelperHomepage}
\alias{inputHelperHomepage}
\title{Set homepage to a hyperlink}
\usage{
inputHelperHomepage(homepage)
}
\arguments{
\item{homepage}{A character string}
}
\value{
A character string or a 'hyperlink' object
}
\description{
Formats homepage as a hyperlink object.
}
