% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_hyperlinks.R
\name{insert_hyperlink}
\alias{insert_hyperlink}
\title{insert_hyperlink()}
\usage{
insert_hyperlink(wb, sheetname, title, sheet_row, index_sheet_name)
}
\arguments{
\item{wb}{Workbook object}

\item{sheetname}{Name of sheet}

\item{title}{Title}

\item{sheet_row}{Sheet row}

\item{index_sheet_name}{Name of sheet where hyperlink should be created}
}
\description{
Function for inserting hyperlinks within an openxlsx Workbook
}
\keyword{internal}
