% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_hyperlinks.R
\name{insert_hyperlinks}
\alias{insert_hyperlinks}
\title{insert_hyperlinks()}
\usage{
insert_hyperlinks(wb, sheetname, title, sheet_row)
}
\arguments{
\item{wb}{Workbook object}

\item{sheetname}{Name of sheet}

\item{title}{Title}

\item{sheet_row}{Sheet row}
}
\description{
Function for inserting hyperlinks within an openxlsx Workbook
}
\keyword{internal}
