% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_metadata_sheet.R
\name{insert_metadata_sheet}
\alias{insert_metadata_sheet}
\title{insert_metadata_sheet()}
\usage{
insert_metadata_sheet(
  wb,
  sheetname = "Metadaten",
  title = "Title",
  source = "statzh",
  metadata = NA,
  logo = "statzh",
  contactdetails = "statzh",
  author = "user",
  date_prefix = "Aktualisiert am: ",
  author_prefix = "durch: ",
  source_prefix = "Datenquelle:",
  metadata_prefix = "Hinweise:"
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{title}{title to be put above the data.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"}

\item{contactdetails}{contact details of the data publisher. Defaults to "statzh".}

\item{author}{defaults to the last two letters (initials) or numbers of the internal user name.}

\item{date_prefix}{Text shown before date}

\item{author_prefix}{Text shown before author name}

\item{source_prefix}{Text shown before sources}

\item{metadata_prefix}{Text shown before metadata}
}
\description{
Function to  add formatted metadata information to an existing .xlsx-workbook.
}
\examples{
# Generation of a spreadsheet
wb <- openxlsx::createWorkbook()
insert_metadata_sheet(wb, title = "Title of mtcars",
  metadata = c("Meta data information."))
\dontrun{
openxlsx::saveWorkbook(wb,"insert_metadata_sheet.xlsx")
}
}
\keyword{insert_metadata_sheet,}
\keyword{openxlsx}
