% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_second_header.R
\name{insert_second_header}
\alias{insert_second_header}
\title{insert_second_header()}
\usage{
insert_second_header(
  wb,
  sheetname,
  data_start_row,
  group_names,
  grouplines,
  data
)
}
\arguments{
\item{wb}{Workbook}

\item{sheetname}{Sheet name}

\item{data_start_row}{Row index for first row with data}

\item{group_names}{Group names}

\item{grouplines}{Group lines}

\item{data}{Data}
}
\description{
Function to export data from R to a formatted .xlsx-file.
}
\note{
The data is exported to the first sheet. Metadata information is exported to
the second sheet.
}
\examples{
\donttest{
\dontrun{
# Beispiel anhand des Datensatzes 'mtcars'
dat <- mtcars

insert_second_header(wb, sheetname, data_start_row, group_names,
 grouplines, data)
}
}
}
\keyword{internal}
