% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_worksheet.R
\name{insert_worksheet}
\alias{insert_worksheet}
\title{worksheetSTAT}
\usage{
insert_worksheet(data, workbook, sheetname, grouplines = FALSE)
}
\arguments{
\item{data}{data to be included in the XLSX-table.}

\item{grouplines}{defaults to FALSE. Can be used to separate grouped variables visually.}

\item{wb}{workbook object to write new worksheet in.}

\item{name}{title of the table / file.}

\item{sheets}{values of the discrete or categorical variable for which a single worksheet should be generated}

\item{sheetvar}{variable which contains the discrete variable to be used as 'period'}
}
\description{
Function to create formatted spreadsheets automatically
}
\examples{
# Generation of a spreadsheet with four worksheets (one per 'carb'-category).
# Can be used to generate worksheets for multiple years.

  # create workbook
wb <- openxlsx::createWorkbook("hello")

worksheeSTAT(mtcars[c(1:10),],wb,"mtcars",c(1:4),carb,grouplines=c(1,5,6))
}
\keyword{worksheetSTAT}
