% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_image.R
\name{insert_worksheet_image}
\alias{insert_worksheet_image}
\title{insert_worksheet_image()}
\usage{
insert_worksheet_image(
  wb,
  sheetname,
  image,
  startrow = 3,
  startcol = 3,
  width,
  height,
  units = "in",
  dpi = 300
)
}
\arguments{
\item{wb}{workbook object to write new worksheet in}

\item{sheetname}{name of the sheet tab}

\item{image}{image or plot}

\item{startrow}{row coordinate of upper left corner of figure}

\item{startcol}{column coordinate of upper left corner of figure}

\item{width}{width of figure}

\item{height}{height of figure}

\item{units}{unit of measurement (default: in)}

\item{dpi}{image resolution (default: 300)}
}
\description{
Insert an image into a worksheet
}
\note{
The function does not write the result into a .xlsx file. A separate
  call to openxlsx::saveWorkbook() is required. A temporary file is created
  for inputs of type gg, ggplot or histogram object at path given by
  `tempfile()`.
}
\examples{
figure <- ggplot2::ggplot(mtcars, ggplot2::aes(x = disp)) +
  ggplot2::geom_histogram()

wb <- openxlsx::createWorkbook()
insert_worksheet_image(wb, sheetname = "ggplot image",
                       image = figure, width = 3.5, height = 5.5)
}
\keyword{insert_worksheet_image}
