% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_worksheet.R
\name{insert_worksheet_nh}
\alias{insert_worksheet_nh}
\title{insert_worksheet_nh}
\usage{
insert_worksheet_nh(
  wb,
  sheetname,
  data,
  title,
  source = getOption("statR_source"),
  metadata = NA,
  grouplines = NA,
  group_names = NA
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{data}{data to be included.}

\item{title}{title to be put above the data.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata information to be included. Defaults to NA.}

\item{grouplines}{Can be used to group variables visually. Values should
either correspond to numeric column indices or column names, denoting the first
variable in a block. Defaults to NA, meaning no lines are added.}

\item{group_names}{Names for groupings in secondary header. Format:
List e.g list(c("title 1", "title 2", "title 3"))}
}
\description{
Function to add formatted worksheets to an existing Workbook.
  The worksheets do not include contact information or logos.
}
\note{
This function doesn't include any contact information. The function
 does not write the result into a .xlsx file. A separate
 call to `openxlsx::saveWorkbook()` is required.
}
\examples{
## Create workbook
wb <- openxlsx::createWorkbook()

## insert a new worksheet
insert_worksheet_nh(wb, sheetname = "data1", data = head(mtcars),
  title = "Title", source = "statzh", metadata = "Note: ...")

## insert a further worksheet
insert_worksheet_nh(wb, sheetname = "data2", data = tail(mtcars),
  title = "Title", source = "statzh", metadata = "Note: ...")

## insert a worksheet with group lines and second header
insert_worksheet_nh(wb, sheetname = "data3", data = head(mtcars),
                    title = "grouplines", source = "statzh",
                    metadata = "Note: ...",
                    grouplines = c(1,5,6), group_names = "carb")
}
\keyword{insert_worksheet_nh}
