% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate2}
\alias{interpolate2}
\title{interpolate2()}
\usage{
interpolate2(palette, color = color, degree = degree, number = number)
}
\arguments{
\item{palette}{a vector of colors to be interpolated}

\item{color}{color (hex-code) to be mixed with the colors in the 'palette'.}

\item{degree}{degree of interpolation, range: 1 (slight interpolation) - 7 (strong interpolation).}

\item{number}{number of colors to be generated.}
}
\description{
Function to generate new color palettes by interpolation.
}
\examples{
#interpolate the 'zhblue'-palette with black and generate 7 new ones
interpolate2(palette = zhpal$zhblue, color = c("#000000"), degree = 3, number = 7)

}
\keyword{interpolate2}
