% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{IRLS}
\alias{IRLS}
\title{Iterative Re-Weighted Least Squares}
\usage{
IRLS(X, y, lam = 0, intercept = TRUE, tol = 10^(-5), maxit = 1e+05, vec)
}
\arguments{
\item{X}{matrix or data frame}

\item{y}{matrix or vector of response 0,1}

\item{lam}{tuning parameter for regularization term}

\item{intercept}{Defaults to TRUE}

\item{tol}{tolerance - used to determine algorithm convergence}

\item{maxit}{maximum iterations}

\item{vec}{optional vector to specify which coefficients will be penalized}

\item{betas}{beta estimates (includes intercept)}
}
\value{
returns beta estimates (includes intercept), total iterations, and gradients.
}
\description{
Computes the logistic regression coefficient estimates using the iterative re-weighted least squares (IRLS) algorithm. This function is to be used with the 'logisticr' function.
}
\examples{
IRLS(X, y, n.list = c(rep(1, n)), lam = 0.1, alpha = 1.5)
}

