% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{is.opts}
\alias{is.opts}
\title{Test whether object is of class "opts"}
\usage{
is.opts(x)
}
\arguments{
\item{x}{The object to test}
}
\value{
A TRUE or FALSE value, indicating whether the object is
of class "opts"
}
\description{
Test whether object is of class "opts"
}
