% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsXLSX_helpers.R
\name{isImplementedPlotType}
\alias{isImplementedPlotType}
\title{isImplementedPlotType()}
\usage{
isImplementedPlotType(object)
}
\value{
A logical value
}
\description{
Check if object is a supported plot object
}
\keyword{internal}
