% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kfold}
\alias{kfold}
\title{K fold (c++)}
\usage{
kfold(n, K)
}
\arguments{
\item{n}{number of eleemtns}

\item{K}{number of folds}
}
\value{
returns vector
}
\description{
creates vector of shuffled indices
}
\examples{
kfold(10, 3)

}
