% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{LASSO}
\alias{LASSO}
\title{Lasso regression}
\usage{
LASSO(X, Y, lam = NULL, intercept = FALSE, standardize = FALSE, ...)
}
\arguments{
\item{X}{nxp data matrix. Each row corresponds to a single observation and each column contains n observations of a single feature/variable.}

\item{Y}{nxr response matrix. Each row corresponds to a single response and each column contains n response of a single feature/response.}

\item{lam}{tuning parameter}

\item{intercept}{option to include intercept, defaults to FALSE}

\item{standardize}{option to standardize the data, defaults to FALSE}

\item{...}{other options to pass to glmnet}
}
\value{
betas, lam
}
\description{
calculate lasso regression coefficients using the optimal tuning parameter from the glmnet package.
}
