% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA}
\alias{LDA}
\title{Linear Discriminant Analysis}
\usage{
LDA(X, y, method = c("MLE", "diagonal", "ridge"), lam = NULL)
}
\arguments{
\item{X}{n x p matrix where the ith row is the values of the predictor for the ith case}

\item{y}{n entry response vector where the ith entry is the response category in {1, ..., C} for the ith case}

\item{method}{estimation method}

\item{lam}{optional tuning parameter specification}
}
\value{
returns a list with the parameter estimates
}
\description{
this function fit the LDA model
}
