% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linearc}
\alias{linearc}
\title{Linearc (c++)}
\usage{
linearc(X, y, lam = 0, weights = 0L, intercept = TRUE, kernel = FALSE)
}
\arguments{
\item{X}{matrix}

\item{y}{matrix}

\item{lam}{optional tuning parameter for ridge regularization term. Defaults to 'lam = 0'}

\item{weights}{optional vector of weights for weighted least squares}

\item{intercept}{add column of ones if not already present. Defaults to TRUE}

\item{kernel}{use linear kernel to compute ridge regression coefficeients. Defaults to true when p >> n}
}
\value{
returns the coefficient estimates
}
\description{
Computes the linear regression coefficient estimates (ridge-penalization and weights, optional)
}
\examples{
Weighted ridge regression
library(dplyr)
X = dplyr::select(iris, -c(Species, Sepal.Length))
y = dplyr::select(iris, Sepal.Length)
linearc(X, y, lam = 0.1, weights = rep(1:150))

Kernelized ridge regression
linearc(X, y, lam = 0.1, kernel = T)

}
