% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{loadDefaultUserConfig}
\alias{loadDefaultUserConfig}
\title{Lade default config}
\usage{
loadDefaultUserConfig(store_path = "~/.config/R/statR")
}
\arguments{
\item{store_path}{Pfad unter welchem das Konfigurationsfile fuer das registrieren
von Header-Config-Yaml-Files liegen soll}
}
\description{
Die Funktion hat nur den Zweck, das Config-File aus dem Paketordner zu laden.
Dadurch wird während der temporären Installation die Möglichkeit geschaffen,
die Konfiguration zu laden, ohne den temporären Pfad im ConfigStore zu
hinterlegen. Die Funktion wid nicht exportiert, weil die Erwartung ist, dass
Benutzer über add_config Konfigurationen registrieren.
}
\keyword{internal}
