% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logitc}
\alias{logitc}
\title{Logitc (c++)}
\usage{
logitc(u)
}
\arguments{
\item{u}{some number}
}
\value{
returns the logit of u
}
\description{
Computes the logit for u
}
\examples{
logit(X*beta)

}
