% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{logitr}
\alias{logitr}
\title{Logit}
\usage{
logitr(u)
}
\arguments{
\item{u}{some number. Ex: X %*% beta}
}
\value{
returns the logit of u
}
\description{
Computes the logit for u
}
\examples{
logit(X \%*\% beta)
}

