% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM.R
\name{MM}
\alias{MM}
\title{Majorize-Minimization function}
\usage{
MM(X, y, lam = 0, alpha = 1.5, gamma = 1, intercept = TRUE,
  tol = 10^(-5), maxit = 1e+05, vec = NULL)
}
\arguments{
\item{X}{matrix or data frame}

\item{y}{matrix or vector of response 0,1}

\item{lam}{optional tuning parameter for ridge regularization term. Defaults to 'lam = 0'}

\item{alpha}{optional tuning parameter for bridge regularization term. Defaults to 'alpha = 1.5'}

\item{gamma}{gamma indicator function. 'gamma = 1' for ridge, 'gamma = 0' for bridge. Defaults to 'gamma = 1'}

\item{intercept}{defaults to TRUE}

\item{tol}{tolerance - used to determine algorithm convergence}

\item{maxit}{maximum iterations}

\item{vec}{optional vector to specify which coefficients will be penalized}
}
\value{
returns beta estimates (includes intercept), total iterations, and gradients.
}
\description{
This function utilizes the MM algorithm. It will be used to compute the logistic regression coefficient estimates. This function is to be used with the 'logisticr' function.
}
\examples{
MM(X, y)
}

