% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MM_linearc}
\alias{MM_linearc}
\title{Linear Majorize-Minimization function (c++)}
\usage{
MM_linearc(X, y, lam = 0, alpha = 1.5, gamma = 1, weights = 0L,
  intercept = TRUE, tol = 1e-05, maxit = 1e+05, vec = 0L, init = 0L)
}
\arguments{
\item{X}{matrix}

\item{y}{matrix}

\item{lam}{optional tuning parameter for ridge regularization term. Defaults to 'lam = 0'}

\item{alpha}{optional tuning parameter for bridge regularization term. Defaults to 'alpha = 1.5'}

\item{gamma}{gamma indicator function. 'gamma = 1' for ridge, 'gamma = 0' for bridge. Defaults to 'gamma = 1'}

\item{intercept}{defaults to TRUE}

\item{tol}{tolerance - used to determine algorithm convergence}

\item{maxit}{maximum iterations}

\item{vec}{optional vector to specify which coefficients will be penalized}

\item{init}{optional initialization for MM algorithm}
}
\value{
returns beta estimates (includes intercept), total iterations, and gradients.
}
\description{
This function utilizes the MM algorithm. It will be used to compute the linear regression coefficient estimates with optional regularization penalties. This function is to be used with the 'linearc' function.
}
\examples{
MM_linearc(X, y)

}
