% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{multiplot}
\alias{multiplot}
\title{Multiple Plot}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{object can be passed in}

\item{plotlist}{plotlist}

\item{cols}{number of columns in layout}

\item{layout}{a matrix specify the layout. If present, 'cols' is ignored}
}
\value{
plots
}
\description{
Taken from: http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
