% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_sum.r
\name{my_sum}
\alias{my_sum}
\title{my_sum: Funktion für einfache (gruppierte) Standradauswertungen
my_sum}
\usage{
my_sum(df, var, ...)
}
\arguments{
\item{df}{dataframe}

\item{var}{variables to be aggregated}
}
\description{
Function to create a formated single-worksheet XLSX automatically
}
\examples{

my_sum(df, var, [grp1, grp2, ...])
df -> data.frame
var -> Auswertungsvariable
grp1 -> Groupingvariable 1 (optional)
grp2 -> Groupingvariable 2 (optional)
}
\keyword{my_sum}
