% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{namedRegionFirstRow}
\alias{namedRegionFirstRow}
\alias{namedRegionLastRow}
\alias{namedRegionFirstCol}
\alias{namedRegionLastCol}
\title{Functions to determine the outer boundaries of named regions}
\usage{
namedRegionFirstRow(wb, sheetname, region_name = NULL)

namedRegionLastRow(wb, sheetname, region_name = NULL)

namedRegionFirstCol(wb, sheetname, region_name = NULL)

namedRegionLastCol(wb, sheetname, region_name = NULL)
}
\arguments{
\item{wb}{A workbook object}

\item{sheetname}{Name of a worksheet}

\item{region_name}{names of regions in Workbook.}
}
\value{
Numeric value
}
\description{
Methods for determining the first and last index of
  a namedRegion in row and column directions.
}
\keyword{internal}
