% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{namedRegionColumnExtent}
\alias{namedRegionColumnExtent}
\title{namedRegionColumnExtent()}
\usage{
namedRegionColumnExtent(wb, sheetname, name = NULL)
}
\arguments{
\item{wb}{A workbook object}
}
\value{
A numeric vector of column indices
}
\description{
Get column extent of a named region in a Workbook object
}
\keyword{internal}
