% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{namedRegionExtent}
\alias{namedRegionExtent}
\title{namedRegionExtent()}
\usage{
namedRegionExtent(wb, sheetname, region_name = NULL, which = "both")
}
\arguments{
\item{wb}{A workbook object}

\item{region_name}{names of regions in Workbook.}

\item{which}{either "row", "col", or "both" (default).}

\item{sheet}{Name of a worksheet}
}
\value{
A list with two numeric vectors row and col, containing
  row and column indices.
}
\description{
Get extent of a named region in a workbook object
}
\details{
If a single name is provided, returns the extent of the associated
  named region. If a character vector is provided, the combined extent is
  returned instead. If left at default (NULL), the combined extent of all
  regions is returned.
}
\keyword{internal}
