% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{namedRegionFirstRow}
\alias{namedRegionFirstRow}
\title{namedRegionFirstRow()}
\usage{
namedRegionFirstRow(wb, sheet, region_name = NULL)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{Name of a worksheet}

\item{region_name}{names of regions in Workbook.}
}
\value{
Numeric value corresponding to first row of named region
}
\description{
Get first row number of named region
}
\keyword{internal}
