% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{namedRegionLastCol}
\alias{namedRegionLastCol}
\title{namedRegionLastCol()}
\usage{
namedRegionLastCol(wb, sheet, region_name = NULL)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{Name of a worksheet}

\item{region_name}{names of regions in Workbook.}
}
\value{
Numeric value corresponding to last column of named region
}
\description{
Get last col number of named region
}
\keyword{internal}
