% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{namedRegionRowExtent}
\alias{namedRegionRowExtent}
\title{namedRegionRowExtent()}
\usage{
namedRegionRowExtent(wb, sheetname, region_name = NULL)
}
\arguments{
\item{wb}{A workbook object}

\item{region_name}{names of regions in Workbook.}
}
\value{
A numeric vector of row indices
}
\description{
Get row extent of a named region in a workbook object
}
\keyword{internal}
