% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_checkr.R
\name{normal_checkr}
\alias{normal_checkr}
\title{Normal Distribution checker}
\usage{
normal_checkr(x)
}
\arguments{
\item{x}{a numerical vector}
}
\description{
Tool to test the distribution of a numerical variable
and checks it against the Normal distribution. THe input
must be a numerical vector and the output is a list.
}
\examples{

normal_checkr(mtcars$mpg)

}
\seealso{
\code{\link{statr}}
}
\author{
Jay Achar
}
