% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{opts}
\alias{opts}
\title{Define options for a statistical function}
\usage{
opts(...)
}
\arguments{
\item{...}{An indefinite number of unquoted options.}
}
\value{
The a character vector of options.
}
\description{
The \code{opts} function is used to define
options for a number of different functions.
}
