% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_linearc}
\alias{predict_linearc}
\title{Predict Linear Regression}
\usage{
predict_linearc(betas, X, y = 0L)
}
\arguments{
\item{betas}{'linearr' object or matrix of betas}

\item{X}{matrix of (new) observations}

\item{y}{matrix of response values}
}
\value{
predictions and loss metrics
}
\description{
Generates prediction for linear regression
}
\examples{

fitted = linearr(X, y, penalty = 'ridge')
predict_linearr(fitted$coefficients, X)

}
