% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_linearr}
\alias{predict_linearr}
\title{Predict Linear Regression}
\usage{
predict_linearr(object, X, y = NULL)
}
\arguments{
\item{object}{'linearr' object or matrix of betas}

\item{X}{matrix or data frame of (new) observations}

\item{y}{optional, matrix or vector of response values}
}
\value{
predictions and loss metrics
}
\description{
Generates prediction for linear regression. Note that one can either input a 'linearr' object or a matrix of beta coefficients.
}
\examples{

fitted = linearr(X, y, lam = 0.1)
predict_linearr(fitted, X)
}

