% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_logisticc}
\alias{predict_logisticc}
\title{Predict Logistic Regression (c++)}
\usage{
predict_logisticc(betas, X, y = 0L)
}
\arguments{
\item{betas}{matrix of coefficientts}

\item{X}{matrix of (new) observations}

\item{y}{matrix of response values 0,1}
}
\value{
predictions and loss metrics
}
\description{
Generates prediction for logistic regression
}
\examples{

fitted = logisticr(X, y, lam = 0.1, penalty = 'ridge', method = 'MM')
predict_logisticr(fitted$coefficients, X)

}
