% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{predict_QDA}
\alias{predict_QDA}
\title{Predict QDA}
\usage{
predict_QDA(fit, Xtest)
}
\arguments{
\item{fit}{this is a list with elements pi.hats, mu.hats, and Sigma.hats where pi.hats is a list of C response category sample proportions, mu.hats is a list of C p-dimensional sample mean proportions, Sigma.hats is a list of C p by p Sample covariance matrices}

\item{Xtest}{this is a matrix with ntest rows and p column, each row is a test case}
}
\value{
returns a vector of ntest entries, where the ith entry is the estimated response category (some value in {1, ..., C}) for the ith test case.
}
\description{
this function classifies test data using a fitted QDA model
}
\examples{
predict_QDA(model, Xtest)
}
