% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{print.exact}
\alias{print.exact}
\title{Print an exact class}
\usage{
\method{print}{exact}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Any follow on parameters.}

\item{verbose}{Whether to print the object in list form.}
}
\value{
The exact object, invisibly
}
\description{
Print an exact class
}
