% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{print.freq}
\alias{print.freq}
\title{Print a freq class}
\usage{
\method{print}{freq}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The freq object to print.}

\item{...}{Any follow on parameters.}

\item{verbose}{Whether to print the object in list form.}
}
\value{
The freq object, invisibly
}
\description{
Print a freq class
}
