% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{QDA}
\alias{QDA}
\title{Quadratic Discriminant Analysis}
\usage{
QDA(X, y, method = c("MLE", "diagonal", "ridge"), lam = NULL)
}
\arguments{
\item{X}{n x p matrix where the ith row is the values of the predictor for the ith case}

\item{y}{n entry response vector where the ith entry is the response category in {1, ..., C} for the ith case}
}
\value{
returns a list with the parameter estimates
}
\description{
this function fit the QDA model
}
\examples{
QDA(X, y, method = 'ridge', lam = seq(0.1, 2, 0.1))
}
