% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_sum.r
\name{quick_sum}
\alias{quick_sum}
\title{quick_sum()}
\usage{
quick_sum(df, var, ..., stats = "all", protect = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{var}{variables to be aggregated}

\item{...}{grouping variables}

\item{stats}{Which descriptive statistics should be computed? Can be "all", "base", or "mean". See details.}

\item{protect}{Apply data protection, can be TRUE or FALSE. See details.}
}
\description{
Function for simple (grouped) descriptive analyses.
}
\details{
If stats is set to "all" (the default), the following descriptive statistics are computed per category of the grouping variable(s): mean, standard deviation,
10\%-quantile, 25\%-quantile, median, 75\%-quantile, 90\%-quantile.
If stats is set to "base", the mean, 25\%-quantile, median, and 75\%-quantile are returned.
stats = "mean" computes the mean and the standard deviation.

If protect is set to TRUE, the mean and the median is only computed for variables/variable categories with four or more observations. For <4 observations, the function returns NA for the respective variable/category. The other descriptives are only computed if there are at least 6 observations per variable (category).
}
\examples{

quick_sum(df=mtcars, var=cyl, mpg, vs, stats="all", protect=FALSE)
}
\keyword{quick_sum}
