% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{readUserConfigStore}
\alias{readUserConfigStore}
\title{Liest das Konfigurationsfile in welchem die Pfade zum Header-Config file
hinterlegt sind}
\usage{
readUserConfigStore(store_path = "~/.config/R/statR")
}
\arguments{
\item{store_path}{Pfad unter welchem das Konfigurationsfile für das registrieren
von Header-Config-Yaml-Files liegen soll}
}
\description{
Liest das Konfigurationsfile in welchem die Pfade zum Header-Config file
hinterlegt sind
}
