% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{removeUserConfig}
\alias{removeUserConfig}
\title{Remove a configuration using a name}
\usage{
removeUserConfig(name, store_path = "~/.config/R/statR")
}
\description{
Remove a configuration using a name
}
