% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{replace_by_parameter}
\alias{replace_by_parameter}
\title{Ersetzt Werte in aktiver Config mit gesetzten Parametern}
\usage{
replace_by_parameter(yaml_file, parameter, config_param_name)
}
\arguments{
\item{yaml_file}{Inhalt der aktiven Config}

\item{parameter}{Wert des Parameters}

\item{config_param_name}{Name des Parameters}
}
\description{
Ersetzt Werte in aktiver Config mit gesetzten Parametern
}
\keyword{internal}
