% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{requests}
\alias{requests}
\title{Define a set of requests for a statistical function}
\usage{
requests(...)
}
\arguments{
\item{...}{An indefinite number of variables and/or variable interactions.}
}
\value{
A character vector of variable names and/or variable interactions.
}
\description{
The \code{requests} function is used to define
requests for the \code{table} parameter on the \code{freq} function.
}
