% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{RIDGE}
\alias{RIDGE}
\title{Ridge regression}
\usage{
RIDGE(X, Y, lam = NULL, intercept = FALSE, standardize = FALSE)
}
\arguments{
\item{X}{nxp data matrix. Each row corresponds to a single observation and each column contains n observations of a single feature/variable.}

\item{Y}{nxr response matrix. Each row corresponds to a single response and each column contains n response of a single feature/response.}

\item{lam}{tuning parameter}
}
\value{
betas, lam
}
\description{
calculate ridge regression coefficients using the optimal tuning parameter from the glmnet package.
}
