% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge2}
\alias{ridge2}
\title{Ridge2}
\usage{
ridge2(X_train, X_test, y_train, y_test, lam.vec)
}
\arguments{
\item{X_train}{training data (matrix or data frame)}

\item{X_test}{validation data (matrix or data frame)}

\item{y_train}{training response}

\item{y_test}{validation reponse}

\item{lam.vec}{vector of lambda values}
}
\value{
returns RSS for each lambda value in lam.vec
}
\description{
This is a special formulation of ridge regression used for cross validation. This function should not be used outside the ridge cross validation function. See: `ridge`.
}

