% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge_CV}
\alias{ridge_CV}
\title{Ridge Cross Validation}
\usage{
ridge_CV(X, y, lam.vec, K = 3)
}
\arguments{
\item{X}{matrix or data frame}

\item{y}{matrix or data frame}

\item{lam.vec}{vector of lambda values}

\item{K}{number of folds in CV (2 or more) - defaults to 3}
}
\value{
returns the coefficient estimates, optimal lambda, and cross validation errors
}
\description{
Cross validation for ridge regression. This function will find the optimal lambda tuning parameter given a vector of optional values.
}
\examples{
library(dplyr)
X = dplyr::select(iris, -c(Species, Sepal.Length))
y = dplyr::select(iris, Sepal.Length)
CV(X, y, lam.vec, K = 5)
}

