% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_stat.R
\name{scale_color_gradient2_stat}
\alias{scale_color_gradient2_stat}
\title{scale_color_gradient2_stat()}
\usage{
scale_color_gradient2_stat(palette = "zhbluered", ..., space = "Lab",
  na.value = "grey50", guide = "colourbar", midpoint = 0)
}
\description{
Stat ZH scales
}
\examples{
scale_color_gradient2_stat()

}
\keyword{scale_color_gradient2_stat}
