% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_stat.R
\name{scale_color_gradientn_stat}
\alias{scale_color_gradientn_stat}
\title{scale_color_gradientn_stat()}
\usage{
scale_color_gradientn_stat(..., colours, values = NULL, space = "Lab",
  na.value = "grey50", guide = "colourbar", colors)
}
\description{
Stat ZH scales
}
\examples{
scale_color_gradientn_stat()

}
\keyword{scale_color_gradientn_stat}
