% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_stat.R
\name{scale_color_stat}
\alias{scale_color_stat}
\title{scale_colour_stat()}
\usage{
scale_color_stat(palette = "zh", ...)
}
\description{
Stat ZH scales
}
\examples{
scale_colour_stat()

}
\keyword{scale_colour_stat}
