% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_stat.R
\name{scale_fill_gradient_stat}
\alias{scale_fill_gradient_stat}
\title{scale_fill_gradient_stat()}
\usage{
scale_fill_gradient_stat(palette = "zhblueseq", ..., space = "Lab",
  na.value = "grey50", guide = "colourbar")
}
\description{
Stat ZH scales
}
\examples{
scale_fill_gradient_stat()

}
\keyword{scale_fill_gradient_stat}
