% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Scatter}
\usage{
scatter(data., x., y.)
}
\arguments{
\item{data.}{data frame}

\item{x.}{x-axis}

\item{y.}{y-axis}
}
\value{
a scatterplot
}
\description{
This function simply streamlines the process of creating a scatterplot with ggplot
}
\examples{
scatter(iris, Sepal.Length, Sepal.Width)
}

