% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{setActiveConfig}
\alias{setActiveConfig}
\title{setActiveConfig()}
\usage{
setActiveConfig(name, persistent = FALSE)
}
\arguments{
\item{name}{Name of the configuration}

\item{persistent}{Whether to load the configuration by default on next
startup}
}
\description{
Sets a configuration to active and optionally makes it
  persistent
}
