% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_bridge.R
\name{softmax}
\alias{softmax}
\title{Softmax}
\usage{
softmax(x, b)
}
\arguments{
\item{x}{data vector}

\item{b}{beta estimates (includes intercept)}
}
\value{
returns the estimated probability
}
\description{
Computes the probability for an individual observation (x) being in class '1'.
}
\examples{
softmax(x, b)
}

