% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_stat.R
\name{stat_div_gradient_pal}
\alias{stat_div_gradient_pal}
\title{stat_div_gradient_pal()}
\usage{
stat_div_gradient_pal(palette = "zhbluered", space = "Lab")
}
\description{
Stat ZH scales
}
\examples{
stat_div_gradient_pal()

}
\keyword{stat_div_gradient_pal}
