% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_stat.R
\name{stat_seq_gradient_pal}
\alias{stat_seq_gradient_pal}
\title{scale_seq_gradient()}
\usage{
stat_seq_gradient_pal(palette = "zhblue", space = "Lab")
}
\description{
Stat ZH scales
}
\examples{
scale_seq_gradient()

}
\keyword{scale_seq_gradient}
