% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statr.R
\docType{package}
\name{statr}
\alias{statr}
\alias{statr-package}
\title{statr: A package of personal statistics tools.}
\description{
The package provides functions to analyse and present data using statistical methods
}
\seealso{
\code{\link{normal_checkr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
