% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_stat.R
\docType{data}
\name{stattheme_data}
\alias{stattheme_data}
\title{Data used by the stat zh palettes}
\format{
A \code{list}.
}
\usage{
stattheme_data
}
\description{
Data used by the stat zh palettes
}
\keyword{datasets}
