% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statXLSX.R
\name{statXLSX}
\alias{statXLSX}
\title{statXLSX}
\usage{
statXLSX(data, name, filename = "file", sheetvar, grouplines = FALSE,
  metadata = NA, logo = "statzh", contactdetails = "statzh",
  source = "statzh")
}
\arguments{
\item{data}{data to be included in the XLSX-table.}

\item{name}{title of the table / file.}

\item{filename}{filename of the xlsx-file. Defaults to "file".}

\item{sheetvar}{variable which contains the variable to be used to split the data across several sheets.}

\item{grouplines}{columns to be separated visually by vertical lines.}

\item{metadata}{metadata-information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"}

\item{contactdetails}{contactdetails of the data publisher. Defaults to "statzh".}

\item{source}{source of the data. Defaults to "statzh".}
}
\description{
Function to create formatted spreadsheets automatically
}
\examples{
# Generation of a spreadsheet with four worksheets (one per 'carb'-category).
# Can be used to generate worksheets for multiple years.

statXLSX(mtcars[c(1:10),],"mtcars",carb,grouplines=c(1,5,6))

statXLSX(head(mtcars),"mpgcars", carb, filename="testfile33",grouplines = c(1,2,3), metadata = "remarks: the source of the data is ...",source="canton of zurich",logo="L:/STAT/08_DS/06_Diffusion/Logos_Bilder/LOGOS/STAT_LOGOS/nacht_map.png")
}
\keyword{statXLSX}
