% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SVDc}
\alias{SVDc}
\title{Linear Singular Value Decomposition (c++)}
\usage{
SVDc(X, y, lam = 0, weights = 0L, intercept = TRUE, kernel = FALSE)
}
\arguments{
\item{X}{matrix}

\item{y}{matrix}

\item{lam}{optional tuning parameter for ridge regularization term. Defaults to 'lam = 0'}

\item{weights}{optional vector of weights for weighted least squares}

\item{intercept}{add column of ones if not already present. Defaults to TRUE}

\item{kernel}{use linear kernel to compute ridge regression coefficeients. Defaults to TRUE when p >> n (for 'SVD')}
}
\value{
returns beta estimates (includes intercept) and gradients.
}
\description{
Computes the logistic regression coefficient estimates using SVD. This function is to be used with the 'linearc' function.
}
\examples{
SVDc(X, y, lam = 0.1 weights = rep(1, 150))

}
