% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_all_variables.R
\name{tabulate_all_variables}
\alias{tabulate_all_variables}
\title{Tabulate data frame summary}
\usage{
tabulate_all_variables(x, unique_value_threshold = 6,
  variable_names = NULL, ...)
}
\arguments{
\item{x}{data frame}

\item{unique_value_threshold}{Integer defining the number of unique 
values above which a variable is analysed as continuous}

\item{variable_names}{Character string to define variable names within output.
Length must be identical to the number of variables.}

\item{...}{Additional arguments to be passed to nested functions (e.g. na.rm = TRUE).}
}
\value{
Data frame of summary values. Counts and percentages for categorical
variables and median and IQR for continuous variables
}
\description{
Tabulate data frame summary
}
\examples{
tabulate_all_variables(iris,
unique_value_threshold = 6,
variable_names = c("Sepal Length", "Sepal Width", "Petal Length",
"Petal Width", "Species"))
}
\seealso{
\code{\link{statr}}
}
\author{
Jay Achar
}
