% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_categorical.R
\name{tabulate_categorical}
\alias{tabulate_categorical}
\title{Tabulate categorical variable}
\usage{
tabulate_categorical(x, round_digits = 1)
}
\arguments{
\item{x}{variable object for tabulation}

\item{round_digits}{integer to define how many digits to round the percentage}
}
\value{
data frame with one row per unique value within the input variable
and columns containing the unique values and their counts with associated percentages
}
\description{
Tabulate categorical variable
}
\examples{
tabulate_categorical(mtcars$vs)
}
\seealso{
\code{\link{statr}}
}
\author{
Jay Achar
}
