% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_continuous.R
\name{tabulate_continuous}
\alias{tabulate_continuous}
\title{Tabulate continuous variable}
\usage{
tabulate_continuous(x, variable_name, round_digits = 1, ...)
}
\arguments{
\item{x}{Variable object for tabulation.}

\item{variable_name}{String variable name to include in first column of output table.}

\item{round_digits}{Integer to define how many digits to round the percentage.}

\item{...}{Additional arguments to be passed to nested functions (e.g. na.rm = TRUE).}
}
\value{
data frame with one row per unique value within the input variable
and columns containing the unique values and their counts with associated percentages
}
\description{
Tabulate continuous variable
}
\examples{
tabulate_continuous(mtcars$mpg, variable_name = "MPG")
}
\seealso{
\code{\link{statr}}
}
\author{
Jay Achar
}
