% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeit.R
\name{timeit}
\alias{timeit}
\title{Time-It}
\usage{
timeit(f)
}
\arguments{
\item{f}{the function to time}
}
\value{
returns the elapsed time
}
\description{
Simple function that prints the computation time of a function
}
\examples{
timeit(lm(dist ~ speed, cars))
}

