% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{tridiag}
\alias{tridiag}
\title{Generate tri-diagonal matrices}
\usage{
tridiag(p = 8, base = 0.7, n = 100, X = FALSE)
}
\arguments{
\item{p}{desired dimension}

\item{base}{base multiplier}
}
\description{
Generate p-dimensional matrices so that its inverse is tri-diagonal.
}
\examples{
tridiag(p = 10, base = 0.7)
}
