% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{verifyInputSheetname}
\alias{verifyInputSheetname}
\alias{verifyInputSheetnames}
\title{Ensure that a sheetname are unique and legal according to MS Excel}
\usage{
verifyInputSheetname(sheetname)

verifyInputSheetnames(sheetnames)
}
\arguments{
\item{sheetname}{A character string with the name for an XLSX worksheet}
}
\value{
A character string
}
\description{
Function which truncates a sheetname to 31 characters
}
\details{
MS Excel imposes a character limit of 31 characters for names of
 worksheets. This function truncates sheetnames accordingly and notify the
 user via a message.
}
\keyword{internal}
