% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{writeUserConfig}
\alias{writeUserConfig}
\title{writeUserConfig()}
\usage{
writeUserConfig(name, config_list)
}
\arguments{
\item{name}{Name of the configuration}

\item{config_list}{List of options set by user}
}
\description{
Writes a user config into a YAML file
}
