% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_additional_text.R
\name{add_additional_text}
\alias{add_additional_text}
\title{add_additional_text()}
\usage{
add_additional_text(wb, text, sheetname, start_row)
}
\arguments{
\item{wb}{workbook object where text should be inserted}

\item{text}{Character vector of text to be inserted}

\item{sheetname}{Sheet where text should be inserted.}

\item{start_row}{Row from which to insert elements}
}
\description{
Insert text into an existing workbook object.
}
\note{
Text is inserted at first column index.
}
\keyword{internal}
