% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{addUserConfig}
\alias{addUserConfig}
\title{Registrieren eines neuen Header-Config-Yaml Files}
\usage{
addUserConfig(name = "default", path = NULL, store_path = "~/.config/R/statR")
}
\arguments{
\item{name}{unter welchem namen soll die Header-Konfiguration abrufbar sein.}

\item{path}{Pfad zum Header-Konfigurations-Yaml-File}
}
\description{
Registrieren eines neuen Header-Config-Yaml Files
}
