% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_xlsx.R
\name{aXLSX}
\alias{aXLSX}
\title{aXLSX()}
\usage{
aXLSX(
  data,
  file,
  title = "Title",
  source = getOption("statR_source"),
  metadata = NA,
  logo = getOption("statR_logo"),
  contactdetails = inputHelperContactInfo(),
  author = "user",
  grouplines = NA,
  group_names = NA
)
}
\arguments{
\item{data}{data to be included.}

\item{file}{Path of output xlsx-file.}

\item{title}{title to be put above the data.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg).
Defaults to "statzh"}

\item{contactdetails}{contact details of the data publisher. Defaults to
"statzh".}

\item{author}{defaults to the last two letters (initials) or numbers of the
internal user name.}

\item{grouplines}{defaults to NA. Can be used to separate grouped
variables visually.}

\item{group_names}{Names for groupings in secondary header. Format:
List e.g list(c("title 1", "title 2", "title 3"))}
}
\description{
Function to export data from R to a formatted .xlsx-file. The
 data is exported to the first sheet. Metadata information is exported to
 the second sheet.
}
\examples{
aXLSX(data = mtcars,
      title = "Motor trend car road tests",
      file = tempfile(fileext = ".xlsx"),
      source = paste("Source: Henderson and Velleman (1981). Building",
                     "multiple regression models interactively.",
                     "Biometrics, 37, 391–411."),
      metadata = paste("The data was extracted from the 1974",
                       "Motor Trend US magazine and comprises fuel",
                       "consumption and 10 aspects of automobile design",
                       "and performance for 32 automobiles",
                       "(1973–74 models)."))
}
\keyword{aXLSX}
