% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsXLSXproto.R
\name{datasetsXLSX}
\alias{datasetsXLSX}
\title{datasetsXLSX}
\usage{
datasetsXLSX(file = "test", datasets = list(head(mtcars), head(diamonds)),
  ...)
}
\arguments{
\item{file}{filename of the xlsx-file. No Default.}

\item{datasets}{data to be included in the XLSX-table.}

\item{name}{title of the table in the worksheet, defaults to "Titel" + the value of the variable used to split the dataset across sheets.}

\item{sheetvar}{variable which contains the variable to be used to split the data across several sheets.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata-information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"}

\item{contactdetails}{contactdetails of the data publisher. Defaults to "statzh".}

\item{grouplines}{columns to be separated visually by vertical lines.}
}
\description{
Function to create formatted spreadsheets automatically
}
\examples{
# Generation of a spreadsheet with four worksheets (one per 'carb'-category).
# Can be used to generate worksheets for multiple years.

datasetsXLSX(file="test",datasets=list(head(mtcars),head(diamonds)))

}
\keyword{datasetsXLSX}
