% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsXLSX.R
\name{datasetsXLSX}
\alias{datasetsXLSX}
\title{datasetsXLSX}
\usage{
datasetsXLSX(file = "test", datasets, sheetnames, titles, sources,
  metadata1, ...)
}
\arguments{
\item{file}{filename of the xlsx-file. No Default.}

\item{datasets}{data to be included in the XLSX-table.}

\item{name}{title of the table in the worksheet, defaults to "Titel" + the value of the variable used to split the dataset across sheets.}

\item{sheetvar}{variable which contains the variable to be used to split the data across several sheets.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata-information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"}

\item{contactdetails}{contactdetails of the data publisher. Defaults to "statzh".}

\item{grouplines}{columns to be separated visually by vertical lines.}
}
\description{
Function to create formatted spreadsheets automatically
}
\examples{
 \donttest{

# example1
datasetsXLSX(file="t8",
           datasets = c(head(mtcars),head(diamonds)),
           sheetnames = c("t1", "t2"),
           titles = c("hi", "hey"),
            sources = c("ji", "hu"),
           metadata1 = c("gut", "schlecht")
)
sql1 <- head(mtcars)
sql4 <- head(diamonds)

datasetsXLSX(file="test",
            datasets = c(sql1, sql2),
           sheetnames = "t1",
           titles = "hey",
           sources = "hu",
           metadata1 = "gut"
) }
}
\keyword{datasetsXLSX}
