% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsXLSX.R
\name{datasetsXLSX}
\alias{datasetsXLSX}
\title{datasetsXLSX}
\usage{
datasetsXLSX(
  file,
  maintitle,
  datasets,
  widths,
  heights,
  startrows,
  startcols,
  sheetnames,
  titles,
  logo = "statzh",
  sources = NULL,
  metadata1 = "",
  auftrag_id,
  ...
)
}
\arguments{
\item{file}{filename of the XLSX file. No Default.}

\item{maintitle}{main title of first sheet "Inhalt".}

\item{datasets}{datasets (dataframes or images) to be included in the XLSX-table.}

\item{widths}{width of figure.}

\item{heights}{height of figure.}

\item{startrows}{row coordinate of upper left corner of figure.}

\item{startcols}{column coordinate of upper left corner of figure.}

\item{sheetnames}{name of the sheet tab.}

\item{titles}{title of the table in the worksheet, defaults to "Titel" + the value of the variable used to split the dataset across sheets.}

\item{sources}{source of the data. Defaults to "statzh".}

\item{metadata1}{metadata-information to be included. Defaults to NA.}

\item{auftrag_id}{order number.}
}
\description{
Function to create formatted XLSX files automatically
}
\examples{

# example with image / plot
# create map

plot <- plot(mtcars$mpg)

datasetsXLSX(file="datasetsXLSX",
            maintitle = "nice datasets",
            datasets = list(head(mtcars),tail(mtcars),plot),
            sheetnames = c("data1","data2","map"),
            widths = c(0,0,5),
            heights = c(0,0,2.5),
            startrows = c(0,0,10),
            startcols = c(0,0,8),
            titles = c("Title","Title", "Plot"),
            sources = c("Quelle: STATENT", "Quelle: Strukturerhebung", "Quelle: Strukturerhebung"),
            metadata1 = c("Bemerkungen: bla", "Bemerkungen: y", "Bemerkungen: x"),
            auftrag_id="A2020_0200"
)


# example without image

datasetsXLSX(file="datasetsXLSX2",
             maintitle = "nice datasets",
             datasets = list(head(mtcars),tail(mtcars)),
             sheetnames = c("data1","data2"),
             titles = c("Title","Title"),
             sources = c("Quelle: STATENT", "Quelle: Strukturerhebung"),
             metadata1 = c("Bemerkungen: y", "Bemerkungen: x"),
             auftrag_id="AS2020_01"
             )


}
\keyword{datasetsXLSX}
