% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsXLSX.R
\name{datasetsXLSX}
\alias{datasetsXLSX}
\title{datasetsXLSX()}
\usage{
datasetsXLSX(
  file,
  maintitle,
  datasets,
  widths = NULL,
  heights = NULL,
  startrows = NULL,
  startcols = NULL,
  sheetnames,
  titles,
  logo = "statzh",
  titlesource = "statzh",
  sources = "statzh",
  metadata1 = NA,
  auftrag_id = NULL,
  contact = "statzh",
  homepage = "statzh",
  openinghours = "statzh"
)
}
\arguments{
\item{file}{file name of the spreadsheet. The extension ".xlsx" is added automatically.}

\item{maintitle}{Title to be put on the first (overview) sheet.}

\item{datasets}{datasets or plots to be included.}

\item{widths}{width of figure in inch (1 inch = 2.54 cm). See details.}

\item{heights}{height of figure in inch (1 inch = 2.54 cm). See details.}

\item{startrows}{row where upper left corner of figure should be placed. See details.}

\item{startcols}{column where upper left corner of figure should be placed. See details.}

\item{sheetnames}{names of the sheet tabs.}

\item{titles}{titles of the different sheets.}

\item{logo}{file path to the logo to be included in the index-sheet. Can be "statzh" or "zh". Defaults to "statzh".}

\item{titlesource}{source to be mentioned on the title sheet beneath the title}

\item{sources}{source of the data. Defaults to "statzh".}

\item{metadata1}{metadata information to be included. Defaults to NA.}

\item{auftrag_id}{order number.}

\item{contact}{contact information on the title sheet. Defaults to "statzh"}

\item{homepage}{web address to be put on the title sheet. Default to "statzh"}

\item{openinghours}{openinghours written on the title sheet. Defaults to Data Shop}
}
\description{
Function to export several datasets and/or figures from R to an .xlsx-file. The function creates an overview sheet and separate sheets
for each dataset/figure.
}
\details{
When including figures, the heights and widths need to be specified as a vector. For example, say you have two datasets and one figure
that you would like to export. widths = c(0,0,5) then suggests that the figure will be 5 inches wide and placed in the third (and last) sheet of the file.
The same logic applies to the specification of `startrows` and `startcols`.
To include a ggplot2-plot, put a print statement around the ggplot2-object: list(..., ..., print(plot)).
}
\examples{
\donttest{
\dontrun{


# Example with two datasets and no figure
dat1 <- mtcars
dat2 <- PlantGrowth

datasetsXLSX(file="twoDatasets", # '.xlsx' wird automatisch hinzugefügt
            maintitle = "Autos und Pflanzen",
            datasets = list(dat1, dat2),
            logo = "statzh",
            sheetnames = c("Autos","Blumen"),
            titles = c("mtcars-Datensatz","PlantGrowth-Datensatz"),
            sources = c("Source: Henderson and Velleman (1981).
            Building multiple regression models interactively. Biometrics, 37, 391–411.",
                        "Dobson, A. J. (1983) An Introduction to Statistical
                        Modelling. London: Chapman and Hall."),
            metadata1 = c("Bemerkungen zum mtcars-Datensatz: x",
                          "Bemerkungen zum PlantGrowth-Datensatz: x"),
            auftrag_id="A2021_0000")

# Example with two datasets and one figure

dat1 <- mtcars
dat2 <- PlantGrowth
fig <- hist(mtcars$disp)

datasetsXLSX(file="twoDatasetsandFigure",
            maintitle = "Autos und Pflanzen", # '.xlsx' wird automatisch hinzugefügt
            datasets = list(dat1, dat2, fig),
            widths = c(0,0,5),
            heights = c(0,0,5),
            startrows = c(0,0,3),
            startcols = c(0,0,3),
            sheetnames = c("Autos","Blumen", "Histogramm"),
            titles = c("mtcars-Datensatz","PlantGrowth-Datensatz", "Histogramm"),
            sources = c("Source: Henderson and Velleman (1981).
            Building multiple regression models interactively. Biometrics, 37, 391–411.",
                        "Source: Dobson, A. J. (1983) An Introduction to
                        Statistical Modelling. London: Chapman and Hall."),
            metadata1 = c("Bemerkungen zum mtcars-Datensatz: x",
                          "Bemerkungen zum PlantGrowth-Datensatz: x"),
            auftrag_id="A2021_0000")
}
}
}
\keyword{datasetsXLSX}
