% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush_left.r
\name{flush_left}
\alias{flush_left}
\title{flush_left}
\usage{
flush_left(g, ...)
}
\arguments{
\item{g}{ggplot object}
}
\description{
Function to flush title, subtitle and caption to the lefthand side of the graphics device
}
\examples{

flush_left(ggplot(mtcars, aes( x = cyl))+
geom_bar()+
labs(title = "Titel", subtitle = "Untertitel", caption = "Datenquelle"))
}
\keyword{flush}
\keyword{left}
