% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groupline_index_by_pattern.R
\name{get_groupline_index_by_pattern}
\alias{get_groupline_index_by_pattern}
\title{get_groupline_index_by_pattern()}
\usage{
get_groupline_index_by_pattern(grouplines, data)
}
\arguments{
\item{grouplines}{defaults to NA. Can be used to separate grouped
variables visually.}

\item{data}{data to be included.}
}
\description{
Derive groupline index by matching names
}
\keyword{internal}
