% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gradient_linearc}
\alias{gradient_linearc}
\title{Gradient of Linear Regression (c++)}
\usage{
gradient_linearc(betas, X, y, lam = 0, weights = 0L, intercept = TRUE)
}
\arguments{
\item{X}{matrix}

\item{y}{response vector of 0,1}

\item{lam}{tuning parameter for ridge regularization term}

\item{weights}{option vector of weights for weighted least squares}

\item{intercept}{add column of ones if not already present. Defaults to TRUE}

\item{beta}{estimates (includes intercept)}
}
\value{
returns the gradient
}
\description{
Computes the gradient of linear regression (optional ridge regularization term). This function is to be used with the 'SVDc' function.
}
\examples{
gradient_linearc(betas, X, y, lam = 0.1, weights = rep(1,150), intercept = TRUE)

}
