% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{initUserConfigStore}
\alias{initUserConfigStore}
\title{Intended to be ran once after setup}
\usage{
initUserConfigStore(path = "~/.config/R/statR")
}
\arguments{
\item{path}{the path to the file which will store names and paths of
different statR configs}
}
\description{
Intended to be ran once after setup
}
