% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_hyperlinks.R
\name{insert_hyperlink}
\alias{insert_hyperlink}
\title{insert_hyperlink()}
\usage{
insert_hyperlink(wb, sheetname, title, sheet_row, index_sheet_name)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{title}{title to be put above the data.}

\item{sheet_row}{Row where hyperlink should be inserted}

\item{index_sheet_name}{Name of sheet where hyperlink should be created}
}
\description{
Function for inserting hyperlinks within an openxlsx Workbook
}
\keyword{internal}
