% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_hyperlinks.R
\name{insert_hyperlinks}
\alias{insert_hyperlinks}
\title{insert_hyperlinks()}
\usage{
insert_hyperlinks(
  wb,
  sheetnames,
  titles,
  index_sheet_name = "Index",
  sheet_start_row = 15
)
}
\arguments{
\item{wb}{Workbook object}

\item{sheetnames}{Names of sheets to create hyperlinks to}

\item{titles}{Titles of Hyperlinks}

\item{index_sheet_name}{Name of sheet where hyperlinks should be created}

\item{sheet_start_row}{Sheet row}
}
\description{
Function for inserting hyperlinks within an openxlsx Workbook
}
\keyword{insert_hyperlinks}
