% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_hyperlinks.R
\name{insert_hyperlinks}
\alias{insert_hyperlinks}
\title{insert_hyperlinks()}
\usage{
insert_hyperlinks(wb, sheetnames, text, where, start_row = 15, file = NULL)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetnames}{Names of sheets to create hyperlinks to}

\item{text}{Text to display in the cell with the hyperlink}

\item{where}{Name of the worksheet where the hyperlinks should be inserted}

\item{start_row}{First row from which hyperlink should be inserted}

\item{file}{External file. Default: NULL}
}
\description{
Function for inserting hyperlinks within an openxlsx Workbook.
             Provides support for links to external .xlsx.
}
\keyword{insert_hyperlinks}
