% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_hyperlinks.R
\name{insert_index_hyperlinks}
\alias{insert_index_hyperlinks}
\title{insert_index_hyperlinks()}
\usage{
insert_index_hyperlinks(
  wb,
  sheetnames,
  titles,
  index_sheet_name = "Index",
  sheet_start_row = 15
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetnames}{Names of sheets to create hyperlinks to}

\item{titles}{Titles of Hyperlinks}

\item{index_sheet_name}{Name of sheet where hyperlink should be created}

\item{sheet_start_row}{Initial row after which hyperlinks should be created}

\item{sheet_row}{Row where hyperlink should be inserted}
}
\description{
Function for inserting hyperlinks within an openxlsx Workbook
}
\keyword{insert_hyperlinks,}
\keyword{internal}
