% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_index_sheet.R
\name{insert_index_sheet}
\alias{insert_index_sheet}
\title{insert_index_sheet()}
\usage{
insert_index_sheet(
  wb,
  sheetname = "Index",
  title,
  auftrag_id,
  logo = getOption("statR_logo"),
  contactdetails = inputHelperContactInfo(),
  homepage = getOption("statR_homepage"),
  openinghours = getOption("statR_openinghours"),
  source = getOption("statR_source")
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{title}{title to be put above the data.}

\item{auftrag_id}{Order ID}

\item{logo}{path of the file to be included as logo (png / jpeg / svg).
Default can be adjusted via user profiles.}

\item{contactdetails}{contact details of the data publisher. Default can be
adjusted via user profiles.}

\item{homepage}{Homepage of data publisher. Default can be adjusted via user
profiles.}

\item{openinghours}{statzh or a character string or vector with opening hours}

\item{source}{source of the data. Default can be adjusted via user profiles}
}
\description{
Function which generates an index sheet inside an openxlsx
  workbook.
}
\details{
The logo can be customized via the `statR_logo` global option. This
  should point to the path of the logo file. Other options such as the image
  size in the final .xlsx can either be changed via the options
  `statR_logo_width` and `statR_logo_height`, or set along with contact
  information in a custom profile.
}
\keyword{insert_index_sheet}
