% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_index_sheet.R
\name{insert_index_sheet}
\alias{insert_index_sheet}
\title{insert_index_sheet()}
\usage{
insert_index_sheet(
  wb,
  sheetname = "Index",
  title,
  auftrag_id,
  logo = "statzh",
  contactdetails = "statzh",
  homepage = "statzh",
  openinghours = "statzh",
  source = "statzh"
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{title}{title to be put above the data.}

\item{auftrag_id}{Order ID}

\item{logo}{path of the file to be included as logo (png / jpeg / svg).
Defaults to "statzh"}

\item{contactdetails}{contact details of the data publisher. Defaults to
"statzh".}

\item{homepage}{Homepage of data publisher. Defaults to "statzh".}

\item{openinghours}{statzh or a character string or vector with opening hours}

\item{source}{source of the data. Defaults to "statzh".}
}
\description{
Function which generates an index sheet inside an openxlsx
  workbook.
}
\keyword{insert_index_sheet}
