% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_index_sheet.R
\name{insert_index_sheet}
\alias{insert_index_sheet}
\title{insert_index_sheet()}
\usage{
insert_index_sheet(
  wb,
  logo = "statzh",
  contact = "statzh",
  homepage = "statzh",
  openinghours = "statzh",
  titlesource,
  auftrag_id,
  maintitle,
  sheetname = "Index",
  date_prefix = "Erstellt am: ",
  order_id_prefix = "Auftragsnr.:"
)
}
\arguments{
\item{wb}{A workbook object as generated by openxlsx::createWorkbook()}

\item{logo}{Either statzh or zh, or path to logo file}

\item{contact}{Either "statzh" or a character string}

\item{homepage}{statzh or a character string with the URL}

\item{openinghours}{statzh or a character string or vector with opening hours}

\item{titlesource}{Character vector denoting the data source}

\item{auftrag_id}{Order ID}

\item{maintitle}{Title shown on index sheet}

\item{sheetname}{Name of Index Sheet}

\item{date_prefix}{Text prepended to date}

\item{order_id_prefix}{Text prepended to order ID}
}
\description{
Function which generates an index sheet inside an openxlsx workbook.
}
\keyword{datasetsXLSX}
