% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_metadata_sheet.R
\name{insert_metadata_sheet}
\alias{insert_metadata_sheet}
\title{insert_metadata_sheet()}
\usage{
insert_metadata_sheet(
  wb,
  sheetname = "Metadaten",
  title = "Title",
  source = getOption("statR_source"),
  metadata = NA,
  logo = getOption("statR_logo"),
  contactdetails = inputHelperContactInfo(compact = TRUE),
  author = "user"
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{title}{title to be put above the data.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg).
Defaults to "statzh"}

\item{contactdetails}{contact details of the data publisher. Defaults to
"statzh".}

\item{author}{defaults to the last two letters (initials) or numbers of the
internal user name.}
}
\description{
Function to add a formatted worksheet with metadata to an
  existing Workbook object.
}
\examples{
# Create Workbook
wb <- openxlsx::createWorkbook()

# Insert a simple metadata sheet
insert_metadata_sheet(wb, title = "Title of mtcars",
  metadata = c("Meta data information."))

}
\seealso{
createWorkbook, addWorksheet, writeData
}
\keyword{insert_metadata_sheet}
