% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_metadata_sheet.R
\name{insert_metadata_sheet}
\alias{insert_metadata_sheet}
\title{insert_metadata_sheet()}
\usage{
insert_metadata_sheet(
  wb,
  sheetname,
  meta_infos,
  logo = getOption("statR_logo"),
  contactdetails = inputHelperContactInfo(compact = TRUE),
  author = "user"
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{meta_infos}{a list with title, source, and metadata as named objects}

\item{logo}{path of the file to be included as logo (png / jpeg / svg).
Default can be adjusted via user profiles.}

\item{contactdetails}{contact details of the data publisher. Default can be
adjusted via user profiles.}

\item{author}{defaults to the last two letters (initials) or numbers of the
internal user name.}
}
\description{
Function to add a formatted worksheet with metadata to an
  existing Workbook object.
}
\seealso{
createWorkbook, addWorksheet, writeData
}
\keyword{insert_metadata_sheet}
