% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_metadata_sheet.R
\name{insert_metadata_sheet}
\alias{insert_metadata_sheet}
\title{insert_metadata_sheet()}
\usage{
insert_metadata_sheet(
  wb,
  sheetname = "Metadaten",
  title = "Title",
  source = "statzh",
  metadata = NA,
  logo = "statzh",
  contactdetails = "statzh",
  author = "user"
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{title}{title to be put above the data.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"}

\item{contactdetails}{contact details of the data publisher. Defaults to "statzh".}

\item{author}{defaults to the last two letters (initials) or numbers of the internal user name.}
}
\description{
Function to  add formatted worksheets to an existing .xlsx-workbook.
}
\examples{

# Generation of a spreadsheet
wb <- openxlsx::createWorkbook("hello")

insert_metadata_sheet(wb, title = "Title of mtcars", metadata = c("Meta data information."))
}
\keyword{insert_metadata_sheet}
