% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_second_header.R
\name{insert_second_header}
\alias{insert_second_header}
\title{insert_second_header()}
\usage{
insert_second_header(
  wb,
  sheetname,
  data_start_row,
  group_names,
  grouplines,
  data
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{data_start_row}{Row index for first row with data}

\item{group_names}{Names for groupings in secondary header. Format:
List e.g list(c("title 1", "title 2", "title 3"))}

\item{grouplines}{defaults to NA. Can be used to separate grouped
variables visually.}

\item{data}{data to be included.}
}
\description{
Function to add a second header row
}
\keyword{internal}
