% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_worksheet.R
\name{insert_worksheet}
\alias{insert_worksheet}
\title{insert_worksheet()}
\usage{
insert_worksheet(
  wb,
  sheetname = "Daten",
  data,
  title,
  source,
  metadata = NA,
  logo = getOption("statR_logo"),
  contactdetails = inputHelperContactInfo(),
  homepage = getOption("statR_homepage"),
  author = "user",
  grouplines = NA,
  group_names = NA
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{data}{data to be included.}

\item{title}{title to be put above the data.}

\item{source}{source of the data. Default can be adjusted via user profiles}

\item{metadata}{metadata information to be included. Defaults to NA, meaning
no metadata are attached.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg).
Default can be adjusted via user profiles.}

\item{contactdetails}{contact details of the data publisher. Default can be
adjusted via user profiles.}

\item{homepage}{Homepage of data publisher. Default can be adjusted via user
profiles.}

\item{author}{defaults to the last two letters (initials) or numbers of the
internal user name.}

\item{grouplines}{Can be used to visually group variables. Values should
either correspond to numeric column indices or column names, denoting the
first variable in a group. Defaults to NA, meaning no lines are added.}

\item{group_names}{A vector of names for groups to be displayed in a
secondary header. Should be of the same length as grouplines, and cannot
be used unless these are set. Defaults to NA, meaning no secondary header
is created.}
}
\description{
Function to insert a formatted worksheet into an existing
 Workbook object.
}
\note{
The function does not write the result into a .xlsx file.
 A separate call to openxlsx::saveWorkbook() is required.
}
\examples{
# Initialize Workbook
wb <- openxlsx::createWorkbook()

# Insert mtcars dataset with STATZH design
insert_worksheet(
  wb = wb, sheetname = "carb", data = mtcars, title = "mtcars dataset")

}
\keyword{insert_worksheet}
