% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_worksheet.R
\name{insert_worksheet}
\alias{insert_worksheet}
\title{insert_worksheet()}
\usage{
insert_worksheet(
  data,
  wb,
  sheetname = "data",
  title = "Title",
  source = "statzh",
  metadata = NA,
  logo = "statzh",
  grouplines = FALSE,
  contactdetails = "statzh",
  homepage = "statzh",
  author = "user",
  date_prefix = "Aktualisiert am: ",
  author_prefix = "durch: "
)
}
\arguments{
\item{data}{data to be included.}

\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{title}{title to be put above the data.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg).
Defaults to "statzh"}

\item{grouplines}{defaults to FALSE. Can be used to separate grouped
variables visually.}

\item{contactdetails}{contact details of the data publisher. Defaults to
"statzh".}

\item{homepage}{Homepage of data publisher. Defaults to "statzh".}

\item{author}{defaults to the last two letters (initials) or numbers of the
internal user name.}

\item{date_prefix}{Text shown before date}

\item{author_prefix}{Text shown before author name}
}
\description{
Function to insert a formatted worksheet into an existing
 Workbook object.
}
\note{
The function does not write the result into a .xlsx file.
 A separate call to openxlsx::saveWorkbook() is required.
}
\examples{
# Initialize an empty Workbook
wb <- openxlsx::createWorkbook()

insert_worksheet(data = head(mtcars),
                 wb = wb,
                 title = "mtcars",
                 sheetname = "carb")

}
\keyword{insert_worksheet}
