% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_worksheet.R
\name{insert_worksheet}
\alias{insert_worksheet}
\title{insert_worksheet()}
\usage{
insert_worksheet(
  wb,
  sheetname = "Daten",
  data,
  title = "Title",
  source = getOption("statR_source"),
  metadata = NA,
  logo = getOption("statR_logo"),
  contactdetails = inputHelperContactInfo(),
  homepage = getOption("statR_homepage"),
  author = "user",
  grouplines = NA,
  group_names = NA
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{data}{data to be included.}

\item{title}{title to be put above the data.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg).
Defaults to "statzh"}

\item{contactdetails}{contact details of the data publisher. Defaults to
"statzh".}

\item{homepage}{Homepage of data publisher. Defaults to "statzh".}

\item{author}{defaults to the last two letters (initials) or numbers of the
internal user name.}

\item{grouplines}{defaults to NA. Can be used to separate grouped
variables visually.}

\item{group_names}{Names for groupings in secondary header. Format:
List e.g list(c("title 1", "title 2", "title 3"))}
}
\description{
Function to insert a formatted worksheet into an existing
 Workbook object.
}
\note{
The function does not write the result into a .xlsx file.
 A separate call to openxlsx::saveWorkbook() is required.
}
\examples{
# Initialize Workbook
wb <- openxlsx::createWorkbook()

# Insert mtcars dataset with STATZH design
insert_worksheet(data = mtcars,
                 wb = wb,
                 title = "mtcars dataset",
                 sheetname = "carb")

}
\keyword{insert_worksheet}
