% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_worksheet.R
\name{insert_worksheet}
\alias{insert_worksheet}
\title{insert_worksheet}
\usage{
insert_worksheet(data, workbook, title = "Title", sheetname = "data",
  source = "statzh", metadata = NA, logo = NULL, grouplines = FALSE,
  contactdetails = "statzh")
}
\arguments{
\item{data}{data to be included in the XLSX-table.}

\item{workbook}{workbook object to write new worksheet in.}

\item{sheetname}{title of the table and the sheet}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata-information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"}

\item{grouplines}{defaults to FALSE. Can be used to separate grouped variables visually.}

\item{contactdetails}{contactdetails of the data publisher. Defaults to "statzh".}
}
\description{
Function to create formatted spreadsheets automatically
}
\examples{
# Generation of a spreadsheet with four worksheets (one per 'carb'-category).
# Can be used to generate worksheets for multiple years.

  # create workbook
wb <- openxlsx::createWorkbook("hello")

insert_worksheet(mtcars[c(1:10),],wb,"mtcars",c(1:4),carb,grouplines=c(1,5,6))
}
\keyword{insert_worksheet}
