% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_image.R
\name{insert_worksheet_image}
\alias{insert_worksheet_image}
\title{insert_worksheet_image()
insert an image into a single worksheet}
\usage{
insert_worksheet_image(
  image,
  wb,
  sheetname,
  startrow = 1,
  startcol = 1,
  width,
  height
)
}
\arguments{
\item{image}{image or plot}

\item{wb}{workbook object to write new worksheet in}

\item{sheetname}{name of the sheet tab}

\item{startrow}{row coordinate of upper left corner of figure}

\item{startcol}{column coordinate of upper left corner of figure}

\item{width}{width of figure}

\item{height}{height of figure}
}
\description{
insert_worksheet_image()
insert an image into a single worksheet
}
\note{
The function does not write the result into a .xlsx file. A separate call
to openxlsx::saveWorkbook() is required.
}
\examples{
library(ggplot2)
wb <- openxlsx::createWorkbook()

# Add a geom_histogram()
#--------------
figure1 <- ggplot(mtcars, aes(x = disp)) +
  geom_histogram()

insert_worksheet_image(image = figure1, wb, sheetname = "ggplot image",
  width = 3.5, height = 5.5)

# Add a base histogram
#--------------
base_hist <- hist(mtcars$disp)

insert_worksheet_image(image = base_hist, wb, sheetname = "Base histogram",
  width = 3.5, height = 5.5)

# Insert an existing image file
#--------------
image_path <- paste0(path.package("statR"), "/extdata/Stempel_Kanton_ZH.png")

insert_worksheet_image(image = image_path, wb, sheetname = "Existing image",
  width = 3.5, height = 5.5)

# Export workbook
#--------------
\dontrun{
openxlsx::saveWorkbook(wb,"insert_worksheet_image.xlsx")
}

}
\keyword{insert_worksheet_image}
