% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_image.R
\name{insert_worksheet_image}
\alias{insert_worksheet_image}
\alias{insert_worksheet_image.default}
\alias{insert_worksheet_image.Content}
\title{Insert image into workbooks}
\usage{
insert_worksheet_image(
  wb,
  sheetname,
  image,
  title = NULL,
  source = NULL,
  metadata = NULL,
  width = NULL,
  height = NULL,
  startrow = 3,
  startcol = 3,
  units = "in",
  dpi = 300
)

\method{insert_worksheet_image}{default}(
  wb,
  sheetname,
  image,
  title = NULL,
  source = NULL,
  metadata = NULL,
  width = NULL,
  height = NULL,
  startrow = 3,
  startcol = 3,
  units = "in",
  dpi = 300
)

\method{insert_worksheet_image}{Content}(
  wb,
  sheetname,
  image,
  title = NULL,
  source = NULL,
  metadata = NULL,
  width = NULL,
  height = NULL,
  startrow = 3,
  startcol = 3,
  units = "in",
  dpi = 300
)
}
\arguments{
\item{wb}{workbook object to write new worksheet in}

\item{sheetname}{Name of the sheet where the image should be inserted}

\item{image}{Image, either a ggplot object or the path to an existing image.}

\item{title}{Title of the image. Can be NULL}

\item{source}{Source associated with the image. Can be NULL.}

\item{metadata}{Metadata associated with the image. Can be NULL.}

\item{width}{width of figure}

\item{height}{height of figure}

\item{startrow}{row coordinate of upper left corner of figure}

\item{startcol}{column coordinate of upper left corner of figure}

\item{units}{unit of measurement (default: in)}

\item{dpi}{image resolution (default: 300)}
}
\description{
Inserts an image into a new worksheet, optionally with title,
  source, and metadata. Images must be provided either as a ggplot object
  or as a path to the image file.
}
\note{
The function does not write the result into a .xlsx file. A separate
  call to openxlsx::saveWorkbook() is required. A temporary file is created
  for inputs of type gg, ggplot or histogram object at path given by
  `tempfile()`.
}
\examples{
figure <- ggplot2::ggplot(mtcars, ggplot2::aes(x = disp)) +
  ggplot2::geom_histogram()

wb <- openxlsx::createWorkbook()
insert_worksheet_image(wb, sheetname = "ggplot image",
                       image = figure, width = 3.5, height = 5.5)
}
\keyword{insert_worksheet_image}
\keyword{internal}
