% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_image.R
\name{insert_worksheet_image}
\alias{insert_worksheet_image}
\title{insert_image}
\usage{
insert_worksheet_image(
  image,
  workbook,
  sheetname,
  startrow,
  startcol,
  width,
  height
)
}
\arguments{
\item{image}{image or plot}

\item{workbook}{workbook object to write new worksheet in}

\item{sheetname}{name of the sheet tab}

\item{startrow}{row coordinate of upper left corner of figure}

\item{startcol}{column coordinate of upper left corner of figure}

\item{width}{width of figure}

\item{height}{height of figure}
}
\value{

}
\description{
insert_image
}
\examples{
# example
\dontrun{
export <- openxlsx::createWorkbook("export")

insert_worksheet_image(image=plot(x = mtcars$wt, y = mtcars$mpg),
                       export,
                       "image",
                       startrow=2,
                       startcol=2,
                       width=3.5,
                       height=5.5
)
openxlsx::saveWorkbook(export,"insert_worksheet_image.xlsx")
}
}
