% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_worksheet_nh.R
\name{insert_worksheet_nh}
\alias{insert_worksheet_nh}
\title{insert_worksheet_nh}
\usage{
insert_worksheet_nh(
  data,
  wb,
  sheetname = "Daten",
  title = "Title",
  source = "statzh",
  metadata = NA,
  grouplines = NA,
  group_names = NA
)
}
\arguments{
\item{data}{data to be included in the XLSX-table.}

\item{wb}{workbook object to write new worksheet in.}

\item{sheetname}{name of the sheet-tab.}

\item{title}{title of the table and the sheet}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata-information to be included. Defaults to NA.}

\item{grouplines}{defaults to NA. Can be used to separate grouped variables visually.}

\item{group_names}{Name(s) of the second header(s).
Format: List e.g list(c("title 1", "title 2", "title 3"))}
}
\description{
Function to add formatted worksheets to an existing Workbook.
  The worksheets do not include contact information or logos.
}
\note{
The function does not write the result into a .xlsx file. A separate
 call to `openxlsx::saveWorkbook()` is required.
}
\examples{
## Create workbook
export <- openxlsx::createWorkbook()

## insert a new worksheet
insert_worksheet_nh(head(mtcars), export, "data1",
  title = "Title", source = "statzh", metadata = "Note: ...")

## insert a further worksheet
insert_worksheet_nh(tail(mtcars), export, "data2",
  title = "Title", source = "statzh", metadata = "Note: ...")

## insert a worksheet with group lines and second header
insert_worksheet_nh(data = head(mtcars), wb = export,
  title = "grouplines", source = "statzh", metadata = "Note: ...",
  grouplines = c(1,5,6), group_names = "carb")

\dontrun{
## save workbook
 openxlsx::saveWorkbook(export,"example.xlsx")
}
}
\keyword{insert_worksheet,}
\keyword{openxlsx}
