% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_worksheet_nh.R
\name{insert_worksheet_nh}
\alias{insert_worksheet_nh}
\title{insert_worksheet_nh}
\usage{
insert_worksheet_nh(
  data,
  wb,
  sheetname = "Daten",
  title = "Title",
  source = "statzh",
  metadata = NA,
  grouplines = NA,
  group_names = NA
)
}
\arguments{
\item{data}{data to be included in the XLSX-table.}

\item{wb}{workbook object to write new worksheet in.}

\item{sheetname}{name of the sheet-tab.}

\item{title}{title of the table and the sheet}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata-information to be included. Defaults to NA.}

\item{grouplines}{defaults to FALSE. Can be used to separate grouped variables visually.}

\item{group_names}{Name(s) of the second header(s). Format: List e.g list(c("title 1", "title 2", "title 3"))}
}
\description{
Function to add formatted worksheets without headers to an existing workbook
}
\examples{

# create workbook
wb <- openxlsx::createWorkbook("hello")

insert_worksheet_nh(mtcars[c(1:10),],wb,"mtcars",c(1:4),"carb",grouplines=c(1,5,6))

# create workbook
export <- openxlsx::createWorkbook("export")

# insert a new worksheet
insert_worksheet_nh(head(mtcars)
                  ,export
                  ,"data1"
                 ,title = "Title"
                  ,source = "Quelle: Statistisches Amt Kanton Z\u00fcrich"
                  ,metadata = "Bemerkung: ...")

 # insert a further worksheet
insert_worksheet_nh(tail(mtcars)
                   ,export
                   ,"data2"
                   ,title = "Title"
                   ,source = "Quelle: Statistisches Amt Kanton Z\u00fcrich"
                   ,metadata = "Bemerkung: ...")
\dontrun{
 # save workbook
 openxlsx::saveWorkbook(export,"example.xlsx")
}
}
\keyword{insert_worksheet}
