% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_worksheet.R
\name{insert_worksheet_nh}
\alias{insert_worksheet_nh}
\title{insert_worksheet_nh}
\usage{
insert_worksheet_nh(
  wb,
  sheetname,
  data,
  title,
  source,
  metadata = NA,
  grouplines = NA,
  group_names = NA
)
}
\arguments{
\item{wb}{workbook object to add new worksheet to.}

\item{sheetname}{name of the sheet tab.}

\item{data}{data to be included.}

\item{title}{title to be put above the data.}

\item{source}{source of the data. Default can be adjusted via user profiles}

\item{metadata}{metadata information to be included. Defaults to NA, meaning
no metadata are attached.}

\item{grouplines}{Can be used to visually group variables. Values should
either correspond to numeric column indices or column names, denoting the
first variable in a group. Defaults to NA, meaning no lines are added.}

\item{group_names}{A vector of names for groups to be displayed in a
secondary header. Should be of the same length as grouplines, and cannot
be used unless these are set. Defaults to NA, meaning no secondary header
is created.}
}
\description{
Function to add formatted worksheets to an existing Workbook.
  The worksheets do not include contact information or logos.
}
\note{
This function doesn't include any contact information. The function
 does not write the result into a .xlsx file. A separate
 call to `openxlsx::saveWorkbook()` is required.
}
\examples{
## Create workbook
wb <- openxlsx::createWorkbook()

## insert a new worksheet
insert_worksheet_nh(wb, sheetname = "data1", data = head(mtcars),
  title = "Title", source = "Source: ...", metadata = "Note: ...")

## insert a further worksheet
insert_worksheet_nh(wb, sheetname = "data2", data = tail(mtcars),
  title = "Title", source = "Source: ...", metadata = "Note: ...")

## insert a worksheet with group lines and second header
insert_worksheet_nh(
  wb, sheetname = "data3", data = head(mtcars), title = "grouplines",
  source = "Source: ...", metadata = "Note: ...",
  grouplines = c(5,8), group_names = c("First group", "Second group"))
}
\keyword{insert_worksheet_nh}
