% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate2}
\alias{interpolate2}
\title{interpolate2()}
\usage{
interpolate2(palette, color = color, degree = degree, number = number)
}
\arguments{
\item{palette}{a vector of colors to be interpolated to the same degree with the same color to generate uniform color palettes}

\item{color}{a single color (hex-code) to be mixed with the colors in the 'palette'.}

\item{degree}{degree of interpolation, range: 1 (slight interpolation) - 7 (strong interpolation).}

\item{number}{number of colours to be generated.}
}
\description{
generate new palettes by interpolation
}
\examples{
#interpolate the 'zhblue'-palette with black and generate 11 new hues.
interpolate2(zhpal$zhblue,"#000000",3,11)

}
\keyword{interpolate2}
