% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{loadUserConfig}
\alias{loadUserConfig}
\title{Setzt die Header-Konfigurationen als options}
\usage{
loadUserConfig(name = "default", store_path = "~/.config/R/statR")
}
\arguments{
\item{name}{welches file soll angezogen werden}

\item{store_path}{Pfad unter welchem das Konfigurationsfile für das registrieren
von Header-Config-Yaml-Files liegen soll}
}
\description{
Setzt die Header-Konfigurationen als options
}
