% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_sum.r
\name{quick_sum}
\alias{quick_sum}
\title{quick_sum}
\usage{
quick_sum(df, var, ..., stats = "all", protect = F)
}
\arguments{
\item{df}{dataframe}

\item{var}{variables to be aggregated}

\item{...}{grouping variables}

\item{stats}{Which evaluation statistics should be displayed (all, base, mean)?}

\item{protect}{Apply data protection T/F}
}
\description{
Function to create a formated single-worksheet XLSX automatically
}
\examples{

quick_sum(df=mtcars, var=cyl, mpg, vs, stats="all", protect=F)
}
\keyword{quick_sum}
