% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickXLSX.R
\name{quickXLSX}
\alias{quickXLSX}
\title{quickXLSX}
\usage{
quickXLSX(data, name, filename = "file", grouplines = 0, metadata = NA,
  logo = statzh, contactdetails = statzhcontact)
}
\arguments{
\item{data}{data to be included in the XLSX-table.}

\item{name}{title of the table / file.}

\item{grouplines}{insert vertical lines to separate columns visually.}
}
\description{
Function to create a formated single-worksheet XLSX automatically
}
\examples{
quickXLSX(data, name, filename="file", grouplines = 0, metadata = NA, logo = statzh,
contactdetails = statzhcontact)
}
\keyword{quickXLSX}
