% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickXLSX.R
\name{quickXLSX}
\alias{quickXLSX}
\title{quickXLSX}
\usage{
quickXLSX(data, name, filename = "file", metadata = NA, logo = "statzh",
  contactdetails = "statzh", source = "statzh", grouplines = FALSE)
}
\arguments{
\item{data}{data to be included in the XLSX-table.}

\item{name}{title of the table.}

\item{filename}{filename of the xlsx-file. Defaults to "file".}

\item{metadata}{metadata-information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"}

\item{contactdetails}{contactdetails of the data publisher. Defaults to "statzh".}

\item{source}{source of the data. Defaults to "statzh".}

\item{grouplines}{columns to be separated visually by vertical lines.}
}
\description{
Function to create a formated single-worksheet XLSX automatically
}
\examples{
quickXLSX(head(mtcars), "mtcars")

#example with own logo, a custom filename, lines separating selected columns and some remarks

quickXLSX(head(mtcars), "mtcars22", filename="testfile22", source="alternative source",contactdetails="blavla", grouplines = c(1,2,3), metadata = "remarks:",logo="L:/STAT/08_DS/06_Diffusion/Logos_Bilder/LOGOS/STAT_LOGOS/nacht_map.png")
}
\keyword{quickXLSX}
