% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickXLSX.R
\name{quickXLSX}
\alias{quickXLSX}
\title{quickXLSX}
\usage{
quickXLSX(data, file, ...)
}
\arguments{
\item{data}{data to be included in the XLSX-table.}

\item{file}{filename of the xlsx-file. No Default.}

\item{title}{title of the table in the worksheet, defaults to "Titel".}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata-information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"}

\item{contactdetails}{contactdetails of the data publisher. Defaults to "statzh".}

\item{grouplines}{columns to be separated visually by vertical lines.}
}
\description{
Function to create a formated single-worksheet XLSX automatically
}
\examples{
quickXLSX(data=head(mtcars), file="mtcars")

#example with own logo, a custom filename, lines separating selected columns and some remarks

quickXLSX(head(mtcars),  file="filename", title="alternative title", source="alternative source",contactdetails="blavla", grouplines = c(1,2,3), metadata = "remarks:",logo="L:/STAT/08_DS/06_Diffusion/Logos_Bilder/LOGOS/STAT_LOGOS/nacht_map.png")
}
\keyword{quickXLSX}
