% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_xlsx.R
\name{quickXLSX}
\alias{quickXLSX}
\title{quickXLSX()}
\usage{
quickXLSX(
  data = NA,
  file,
  title = "Title",
  source = getOption("statR_source"),
  metadata = NA,
  logo = getOption("statR_logo"),
  contactdetails = inputHelperContactInfo(compact = TRUE),
  author = "user",
  grouplines = NA,
  group_names = NA
)
}
\arguments{
\item{data}{data to be exported.}

\item{file}{file name of the xlsx-file. The extension ".xlsx" is added}

\item{title}{title to be put above the data.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg).
Defaults to "statzh"}

\item{contactdetails}{contact details of the data publisher. Defaults to
"statzh".}

\item{author}{defaults to the last two letters (initials) or numbers of the
internal user name.}

\item{grouplines}{Can be used to group variables visually. Values should
either correspond to numeric column indices or column names, denoting the first
variable in a block. Defaults to NA, meaning no lines are added.}

\item{group_names}{Names for groupings in secondary header. Format:
List e.g list(c("title 1", "title 2", "title 3"))}
}
\description{
A simple function for exporting data from R to a single formatted
  .xlsx-spreadsheet.
}
\examples{

title <- "Motor trend car road tests"
source <- paste(
  "Henderson and Velleman (1981). Building",
  "multiple regression models interactively.",
  "Biometrics, 37, 391–411.")
metadata <- paste(
  "The data was extracted from the 1974 Motor",
  "Trend US magazine and comprises fuel consumption",
  "and 10 aspects of automobile design and",
  "performance for 32 automobiles (1973–74 models).")

quickXLSX(data = mtcars,
          file = tempfile(fileext = ".xlsx"),
          title = title,
          source = source,
          metadata = metadata)

}
\keyword{quickXLSX}
