% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{readUserConfig}
\alias{readUserConfig}
\title{Liest das Header-Konfigurations-Yaml-File}
\usage{
readUserConfig(name = "default", store_path = "~/.config/R/statR")
}
\arguments{
\item{name}{welches file soll angezogen werden}
}
\description{
Liest das Header-Konfigurations-Yaml-File
}
