% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{readUserConfig}
\alias{readUserConfig}
\title{readUserConfig()}
\usage{
readUserConfig(name = "default")
}
\arguments{
\item{name}{The name of the configuration}
}
\description{
Reads a user config. By default loads the default config
}
