% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{removeUserConfig}
\alias{removeUserConfig}
\title{Loeschen eines Konfigurations-Eintrages}
\usage{
removeUserConfig(name, store_path = "~/.config/R/statR")
}
\arguments{
\item{name}{Welcher registrierte Eintrag soll geloescht werden}

\item{store_path}{Pfad unter welchem das Konfigurationsfile fuer das registrieren
von Header-Config-Yaml-Files liegen soll}
}
\description{
Loeschen eines Konfigurations-Eintrages
}
