% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{removeUserConfig}
\alias{removeUserConfig}
\title{Löschen eines Konfigurations-Eintrages}
\usage{
removeUserConfig(name, store_path = "~/.config/R/statR")
}
\arguments{
\item{name}{Welcher registrierte Eintrag soll gelöscht werden}
}
\description{
Löschen eines Konfigurations-Eintrages
}
