% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Ridge Regression}
\usage{
ridge(X, y, lam, K = NULL)
}
\arguments{
\item{X}{matrix or data frame}

\item{y}{matrix or data frame of response values}

\item{lam}{tuning parameter for regularization term}

\item{K}{optional K-fold cross validation}
}
\value{
returns the coefficient estimates
}
\description{
Computes the ridge (L2)-penalized linear regression estimates
}
\examples{
library(dplyr)
X = dplyr::select(iris, -c(Species, Sepal.Length))
y = dplyr::select(iris, Sepal.Length)
Ridge(X, y, 0.1)
}

