% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitXLSX.R
\name{splitXLSX}
\alias{splitXLSX}
\title{splitXLSX}
\usage{
splitXLSX(
  data,
  file,
  sheetvar,
  title = "Titel",
  source = "statzh",
  metadata = NA,
  logo = NULL,
  grouplines = FALSE,
  contactdetails = "statzh",
  author = "user"
)
}
\arguments{
\item{data}{data to be included in the XLSX-table.}

\item{file}{filename of the xlsx-file. No Default.}

\item{sheetvar}{variable which contains the variable to be used to split the data across several sheets.}

\item{title}{title of the table in the worksheet, defaults to "Titel" + the value of the variable used to split the dataset across sheets.}

\item{source}{source of the data. Defaults to "statzh".}

\item{metadata}{metadata-information to be included. Defaults to NA.}

\item{logo}{path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"}

\item{grouplines}{columns to be separated visually by vertical lines.}

\item{contactdetails}{contactdetails of the data publisher. Defaults to "statzh".}

\item{author}{defaults to last two letters (initials) or numbers of the user name.}
}
\description{
Function to create formatted spreadsheets with multiple worksheets automatically.
}
\examples{
# Generation of a spreadsheet with four worksheets (one per 'carb'-category).
# Can be used to generate worksheets for multiple years.

splitXLSX(head(mtcars),"mtcars",carb,grouplines=c(1,5,6))

splitXLSX(head(mtcars),
sheetvar=carb,
title="Tabelle",
file="filename",
grouplines = c(1,2,3),
metadata = "remarks: ....",
source="canton of zurich",
logo="L:/STAT/08_DS/06_Diffusion/Logos_Bilder/LOGOS/STAT_LOGOS/nacht_map.png")
}
\keyword{splitXLSX}
