% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statXLSX.R
\name{statXLSX}
\alias{statXLSX}
\title{statXLSX}
\usage{
statXLSX(data, name, sheets, sheetvar, grouplines = FALSE)
}
\arguments{
\item{data}{data to be included in the XLSX-table.}

\item{name}{title of the table / file.}

\item{sheets}{values of the discrete or categorical variable for which a single worksheet should be generated}

\item{sheetvar}{variable which contains the discrete variable to be used as 'period'}

\item{grouplines}{defaults to FALSE. Can be used to separate grouped variables visually.}
}
\description{
Function to create formatted spreadsheets automatically
}
\examples{
# Generation of a spreadsheet with four worksheets (one per 'carb'-category). 
# Can be used to generate worksheets for multiple years.

statXLSX(mtcars[c(1:10),],"mtcars",c(1:4),carb,grouplines=c(1,5,6))
}
\keyword{statXLSX}
