% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{updateUserConfig}
\alias{updateUserConfig}
\title{Anpassen eines Header-Config-Yaml-Files-Pfades}
\usage{
updateUserConfig(name, path, store_path = "~/.config/R/statR")
}
\arguments{
\item{name}{Zu welchem Eintrag moechtest du den Pfad anpassen}

\item{path}{Pfad zum Header-Konfigurations-Yaml-File}

\item{store_path}{Pfad unter welchem das Konfigurationsfile fuer das registrieren
von Header-Config-Yaml-Files liegen soll}
}
\description{
Anpassen eines Header-Config-Yaml-Files-Pfades
}
