% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeText.R
\name{writeText}
\alias{writeText}
\title{Compact function to write text into a worksheet and apply a style}
\usage{
writeText(wb, sheetname, x, row, column, style, name)
}
\arguments{
\item{wb}{A workbook object}

\item{sheetname}{A sheetname}

\item{x}{Input}

\item{row}{The row at which to insert text}

\item{column}{The column at which to insert text}

\item{style}{A style object as defined by openxlsx::createStyle}

\item{name}{A name to assign to the text. Note that this has to be unique}
}
\description{
A shorthand function for openxlsx::writeData and openxlsx::addStyle
}
\keyword{writeText}
